An R6 Class for an edibble design

An object with EdibbleDesign holds the core information related to the experimental design.

Active bindings

active

get the active view

name

The name of the design or project

names_trts

The treatment names.

names_units

The unit variable names.

names_resp

The unit variable names.

names_vars

The name of all variables.

graph

An edibble graph.

table

A table served from the edibble graph.

validation

Data validation for response.

context

The context of the experiment.

Methods

Public methods

EdibbleDesign$new()

EdibbleDesign$finalize()

EdibbleDesign$plot()

EdibbleDesign$print()

EdibbleDesign$save_seed()

EdibbleDesign$muffle()

EdibbleDesign$chatty()

EdibbleDesign$activate_table()

EdibbleDesign$activate_graph()

EdibbleDesign$rename()

EdibbleDesign$add_variable()

EdibbleDesign$add_variable_node()

EdibbleDesign$add_allocation()

EdibbleDesign$assign_allocation()

EdibbleDesign$append_validation()

EdibbleDesign$delete_variable()

EdibbleDesign$clone()

Method new()

Initialise the edibble design. See also start_design().

Usage

EdibbleDesign$new(name = NULL)

Arguments

name

The name or title of the project. The title should be informative and not overly long. If not supplied the default is "An edibble design".

Method finalize()

Invoked when the EdibbleDesign object is deleted.

Usage

EdibbleDesign$finalize()

Method plot()

Plot the intermediate contruct.

Usage

EdibbleDesign$plot(view = c("high", "low"), ..., main = NULL)

Arguments

view

A high- or -low level view.

...

Arguments passed into plot.igraph.

main

The title of the plot. By default it is the name of the EdibbleDesign object.

Method print()

Print the intermediate construct.

Usage

EdibbleDesign$print(...)

Arguments

...

Arguments passed into print.

Method save_seed()

This saves the .Random.seed used just prior to the randomisation process.

Usage

EdibbleDesign$save_seed()

Method muffle()

Turn off messages.

Usage

EdibbleDesign$muffle()

Method chatty()

Turn on messages.

Usage

EdibbleDesign$chatty()

Method activate_table()

Change active view to table.

Usage

EdibbleDesign$activate_table()

Method activate_graph()

Change active view to graph.

Usage

EdibbleDesign$activate_graph()

Method rename()

Rename the variables

Usage

EdibbleDesign$rename(...)

Arguments

...

Similar to dplyr::rename, use new_name = old_name to rename variables.

Method add_variable()

Add the variable to the graph.

Usage

EdibbleDesign$add_variable(value, vname, attr)

Arguments

value

Short hand value of what to add.

vname

The variable name

attr

The vertex attributes.

Method add_variable_node()

Add the variable to the graph.

Usage

EdibbleDesign$add_variable_node(vnames_new, vname_unit, attr)

Arguments

vnames_new

The names of new variables to add.

vname_unit

The unit variable that measurement is taken on.

attr

The vertex attributes.

Method add_allocation()

This adds edges that connect treatment to unit.

Usage

EdibbleDesign$add_allocation(trts, unit)

Arguments

trts

The name of the treatments.

unit

The name of the unit. There should be only one unit.

Method assign_allocation()

This removes "t2vmay" edges and assigns edges from treatment to variable.

Usage

EdibbleDesign$assign_allocation()

Method append_validation()

Appending validation rule

Usage

EdibbleDesign$append_validation(validation)

Arguments

validation

A new list of validation. The name of the list should correspond to a record variable. The elements of the list should be a named list with "operator", "value" and "type". If type is list then it is will contain only "type" and "values".

Method delete_variable()

Delete variable

Usage

EdibbleDesign$delete_variable(vnames)

Arguments

vnames

The name of the variable to delete.

Method clone()

The objects of this class are cloneable with this method.

Usage

EdibbleDesign$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.