An R6 Class for a named experimental design

An R6 Class for a named experimental design

An R6 Class for a named experimental design

Public fields

name

The short name of the experimental design.

name_full

The full name of the experimental design.

code

The code, based on the grammar of experimental design, to generate the experimental design.

n

The total number of experimental units.

t

The total number of treatments tested.

info_always

A character vector containing information or properties that is always true for this named design. Each entry in the vector forms a dot point.

info_this

A character vector containing information or properties that is true for this named design for the chosen parameters.

Methods

Public methods

NamedDesign$new()

NamedDesign$add_code()

NamedDesign$add_seed()

NamedDesign$add_info()

NamedDesign$final_code()

NamedDesign$decorate_units()

NamedDesign$decorate_trts()

NamedDesign$clone()

Method new()

Initialise the named design.

Usage

NamedDesign$new(name, name_full = name, n, t)

Arguments

name

Short name.

name_full

The full name of the design.

n

Total number of experimental units.

t

Total number of treatments tested.

seed

The seed number.

Method add_code()

Append code.

Usage

NamedDesign$add_code(x)

Arguments

x

A string specifying the code to append.

Method add_seed()

Add code to set seed at the beginning.

Usage

NamedDesign$add_seed(seed)

Arguments

seed

The seed number.

Method add_info()

Add information about the named design.

Usage

NamedDesign$add_info(x, always = FALSE)

Arguments

x

A string specifying one info about a design.

always

A logical to specify whether the added information is always true for this this design or only for the given parameter. Default is FALSE.

Method final_code()

This adds the final bit of code needed to generate the edibble data frame.

Usage

NamedDesign$final_code()

Method decorate_units()

A function to decorate units.

Usage

NamedDesign$decorate_units(x)

Arguments

x

A string to decorate.

Method decorate_trts()

A function to decorate treatments.

Usage

NamedDesign$decorate_trts(x)

Arguments

x

A string to decorate.

Method clone()

The objects of this class are cloneable with this method.

Usage

NamedDesign$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.