Activate the provenance in the edibble design object

This is a developer function to create a new Kitchen class with the existing design.

Usage

activate_provenance(
  .edibble,
  overwrite = c("graph", "anatomy", "recipe", "validation", "simulate")
)

Arguments

.edibble

An edibble object.

overwrite

What object to overwrite in the provenance object.

Value

A Provenance object.

Examples

activate_provenance(takeout())
#> → No name was supplied so selecting a random named experimental design...
#> → Selected Strip-Plot DesignSelected Strip-Unit Design
#> <Provenance>
#>   Public:
#>     append_fct_edges: function (from, to, type = NULL, group = NULL, attrs = NULL) 
#>     append_fct_nodes: function (name, role, attrs = NULL) 
#>     append_lvl_edges: function (from, to, attrs = NULL) 
#>     append_lvl_nodes: function (value, n = NULL, label = NULL, attrs = NULL, fid = NULL) 
#>     clone: function (deep = FALSE) 
#>     deactivate: function (delete = c("graph", "anatomy", "recipe", "validation")) 
#>     fct_edges: active binding
#>     fct_exists: function (id = NULL, name = NULL, role = NULL, abort = TRUE) 
#>     fct_graph_components: function (id = NULL) 
#>     fct_id: function (name = NULL, role = NULL) 
#>     fct_id_ancestor: function (id = NULL, role = NULL) 
#>     fct_id_child: function (id = NULL, role = NULL) 
#>     fct_id_descendant: function (id = NULL, role = NULL) 
#>     fct_id_from_lvl_id: function (id = NULL, fid_search = NULL) 
#>     fct_id_from_lvl_values: function (value = NULL, fid_search = NULL) 
#>     fct_id_leaves: function (role = NULL) 
#>     fct_id_links: function (id = NULL, role = NULL, link = c("direct", "indirect")) 
#>     fct_id_parent: function (id = NULL, role = NULL, type = NULL) 
#>     fct_levels: function (id = NULL, name = NULL, return = c("id", "value")) 
#>     fct_levels_id_to_edbl_fct: function (fct_levels, role) 
#>     fct_levels_id_to_value: function (fct_levels) 
#>     fct_levels_value_to_id: function (fct_levels) 
#>     fct_n: active binding
#>     fct_names: function (id = NULL, role = NULL) 
#>     fct_nodes: active binding
#>     fct_role: function (id = NULL) 
#>     get_edibble_version: function () 
#>     get_graph: function () 
#>     get_seed: function () 
#>     get_session_info: function () 
#>     get_simulate: function (name = NULL) 
#>     get_simulate_result_env: function (name = NULL) 
#>     get_title: function () 
#>     get_trail: function () 
#>     get_validation: function (type = NULL) 
#>     graph_subset: function (id = NULL, include = c("self", "child", "parent", "ancestors")) 
#>     initialize: function (graph = NULL) 
#>     is_connected: active binding
#>     lvl_edges: active binding
#>     lvl_graph_components: function () 
#>     lvl_id: function (value = NULL, role = NULL, fid = NULL) 
#>     lvl_id_ancestor: function (id = NULL, role = NULL) 
#>     lvl_id_child: function (id = NULL, role = NULL) 
#>     lvl_id_from_fct_id: function (fid = NULL) 
#>     lvl_id_parent: function (id = NULL, role = NULL) 
#>     lvl_mapping: function (from, to, return = c("vector", "table")) 
#>     lvl_n: active binding
#>     lvl_nodes: active binding
#>     lvl_values: function (id = NULL, role = NULL, fid = NULL) 
#>     make_trts_table: function (id = NULL, return = c("id", "value")) 
#>     mapping: function (role_from, role_to) 
#>     mapping_to_unit: function (id = NULL) 
#>     rcrd_class: function (name = NULL) 
#>     rcrd_exists: function (id = NULL, name = NULL, abort = TRUE) 
#>     rcrd_ids: active binding
#>     rcrd_names: function (id = NULL) 
#>     rcrd_values: function (uid = NULL, fid = NULL) 
#>     reactivate: function (design, overwrite = c("graph", "anatomy", "recipe", 
#>     record_step: function () 
#>     record_track_external: function (code) 
#>     save_seed: function (seed, type) 
#>     serve_rcrds: function (id = NULL, return = c("id", "value")) 
#>     serve_trts: function (id = NULL, return = c("id", "value")) 
#>     serve_units: function (id = NULL, return = c("id", "value")) 
#>     set_name: function (name) 
#>     set_simulate: function (name, process, rcrds) 
#>     set_title: function (title) 
#>     set_validation: function (validation, type = "rcrds") 
#>     trt_exists: function (id = NULL, name = NULL, abort = TRUE) 
#>     trt_ids: active binding
#>     trt_names: function (id = NULL) 
#>     trt_values: function (id = NULL, fid = NULL) 
#>     unit_exists: function (id = NULL, name = NULL, abort = TRUE) 
#>     unit_ids: active binding
#>     unit_names: function (id = NULL) 
#>     unit_values: function (id = NULL, fid = NULL) 
#>   Private:
#>     add_trail_internal: function (code) 
#>     anatomy: formula
#>     build_condtable: function (subgraph, return) 
#>     build_subtable: function (subgraph, return) 
#>     edbl_version: package_version, numeric_version
#>     fct_id_last: 6
#>     fct_new_id: function (n = 1) 
#>     graph: edbl_graph
#>     graph_topological_order: function (graph, reverse = TRUE) 
#>     lvl_id_last: 49
#>     lvl_new_id: function (n = 1) 
#>     name: edibble
#>     node_id_parent_child: function (id = NULL, role = NULL, type = NULL, node = c("factor", 
#>     recipe: design(title = "Strip-Plot Design | Strip-Unit Design")  ...
#>     record_track_internal: function () 
#>     seed: list
#>     session_info: sessionInfo
#>     simulate: list
#>     simulate_result_env: environment
#>     table: list
#>     title: Strip-Plot Design | Strip-Unit Design
#>     trail: list
#>     validate_id: function (id, n = NULL, role = NULL) 
#>     validate_name: function (name) 
#>     validate_role: function (role) 
#>     validation: list
#>     var_id_ancestor: function (id = NULL, role = NULL, node = c("factor", "level")) 
#>     var_id_descendant: function (id = NULL, role = NULL, node = c("factor", "level"))