Cook the design in the kitchen

This is a developer function to create a new Kitchen class with the existing design.

Usage

cook_design(x)

Arguments

x

An edibble object.

Value

A Kitchen object.

Examples

cook_design(takeout())
#> → No name was supplied so selecting a random named experimental design...
#> → Selected Hyper-Graeco-Latin Square Design
#> <Kitchen>
#>   Public:
#>     add_anatomy: function (fresh, name, class) 
#>     append_fct_edges: function (data) 
#>     append_fct_nodes: function (data) 
#>     append_lvl_edges: function (data) 
#>     append_lvl_nodes: function (data) 
#>     clone: function (deep = FALSE) 
#>     design: edbl_design, edbl
#>     fct_ancestor: function (id = NULL, class = NULL) 
#>     fct_child: function (id = NULL, class = NULL) 
#>     fct_class: function (id = NULL) 
#>     fct_edges: active binding
#>     fct_exists: function (name = NULL, id = NULL, class = NULL, abort = TRUE) 
#>     fct_id: function (name = NULL, class = NULL) 
#>     fct_last_id: active binding
#>     fct_leaves: active binding
#>     fct_levels: function (id = NULL, name = NULL) 
#>     fct_n: active binding
#>     fct_names: function (id = NULL, class = NULL) 
#>     fct_nodes: active binding
#>     fct_parent: function (id = NULL, class = NULL) 
#>     initialize: function (design = NULL) 
#>     is_connected: active binding
#>     lvl_ancestor: function (id = NULL, class = NULL) 
#>     lvl_child: function (id = NULL, class = NULL) 
#>     lvl_class: function (id = NULL) 
#>     lvl_edges: active binding
#>     lvl_id: function (name = NULL, class = NULL) 
#>     lvl_last_id: active binding
#>     lvl_n: active binding
#>     lvl_names: function (id = NULL, class = NULL) 
#>     lvl_nodes: active binding
#>     lvl_parent: function (id = NULL, class = NULL) 
#>     rcrd_ids: active binding
#>     rcrd_names: active binding
#>     rcrds_exists: function (abort = TRUE) 
#>     setup_data: function (fresh, name, class) 
#>     trt_ids: active binding
#>     trt_names: active binding
#>     trts_exists: function (abort = TRUE) 
#>     unit_ids: active binding
#>     unit_names: active binding
#>     units_exists: function (abort = TRUE) 
#>   Private:
#>     fresh_type: function (fresh) 
#>     next_method: function (generic, class) 
#>     setup_data.cross_lvls: function (fresh, name, class) 
#>     setup_data.default: function (fresh, name, class) 
#>     setup_data.edbl_fct: function (fresh, name, class) 
#>     setup_data.edbl_lvls: function (fresh, name, class) 
#>     setup_data.formula: function (fresh, name, class) 
#>     setup_data.nest_lvls: function (fresh, name, class) 
#>     var_names: function (nodes, id, class) 
#>     version: NULL