edibble: Designing Comparative Experiments

A system to facilitate designing comparative experiments using the grammar of experimental designs https://emitanaka.org/edibble-book/. An experimental design is treated as an intermediate, mutable object that is built progressively by fundamental experimental components like units, treatments, and their relation.

Details

(WIP)

Website

The website for the package is at https://edibble.emitanaka.org

Discussion is at https://github.com/emitanaka/edibble/discussions

Package options

The following options are used for changing the default view for the print out of edibble design or edibble graph.

edibble.tree.decorate.trts

edibble.tree.decorate.units

edibble.tree.decorate.rcrd

edibble.tree.decorate.levels

edibble.tree.decorate.main

TODO

See also

Author

Maintainer: Emi Tanaka dr.emi.tanaka@gmail.com (ORCID) [copyright holder]