Lady tasting tea

Lady tasting tea experiment was described in Fisher (1935) to test the ability of a lady who said she tell whether the tea or milk was added first to a cup of tea.

The experiment consisted of preparing eight cups of tea, four with milk poured first and the other four with tea poured first. The lady has been told in advance that there are four of each kind of preparation.

This data consists of the same experimental structure and result but the order presented in practice is unknown.

cup

The cup number.

first

The cup of tea prepared with milk or tea first.

guess

The guess by lady which one was poured first.

correct

Whether the lady's guess was correct.

Usage

lady_tasting_tea

Format

An object of class tbl_df (inherits from tbl, data.frame) with 8 rows and 4 columns.

Source

Fisher, Ronald (1935) The Design of Experiments.

See also