Create a classical named experimental design

The function make_classical generates a classical named experimental design by supplying its short name and prints out, by default:

info: information about the named experimental design,

code: code to create the design using edibble, and

table: an edibble data frame for the generated design.

You can find the available short names with find_classical_names().

make_classical(.name = "", ..., .seed = sample(1000, 1), .output = TRUE)

code_classical(.name = "", .seed = sample(1000, 1), ..., .quiet = FALSE)

Arguments

.name
The short name of the classical named experimental design. See under Details for the available named designs.

...
Parameters passed into the prep_classical_* functions.

.output
A logical value to indicate whether all output should be printed or not or a vector of character (e.g. c("info", "code", "table")) specifying which of the three outputs should be printed. Default is TRUE.

.quiet
Opposite of .code. Whether to suppress code output.

Details

For The available named designs are:

"crd": completely randomised design

"rcbd": randomised complete block design

"split": split plot design

See also

Examples

make_classical("crd", n = 50, t = 5)
#> 
#> ── experimental design details ─────────────────────────────────────────────────
#>   • This experimental design is often called Completely Randomised Design.
#>   • You can change the number in `set.seed` to get another random instance of
#>   the same design.
#>   • This design has a total of 50 units testing a total of 5 treatments.
#> The following information is only true for the chosen parameters and not
#> necessary true for all Completely Randomised Designs.
#>   • This design is balanced for the given numbers.
#> 
#> ── edibble code ────────────────────────────────────────────────────────────────
#> set.seed(173)
#> start_design("crd") %>%
#>   set_units(unit = 50) %>%
#>   set_trts(treat = 5) %>%
#>   allot_trts(treat ~ unit) %>%
#>   assign_trts("random") %>%
#>   serve_table() 
#> 
#> ── edibble data frame ──────────────────────────────────────────────────────────
#> # An edibble: 50 x 2
#>          unit    treat
#>    <unit(50)> <trt(5)>
#>  1     unit1    treat3
#>  2     unit2    treat1
#>  3     unit3    treat2
#>  4     unit4    treat4
#>  5     unit5    treat1
#>  6     unit6    treat4
#>  7     unit7    treat5
#>  8     unit8    treat3
#>  9     unit9    treat2
#> 10     unit10   treat1
#> # … with 40 more rows
# if you omit the design parameters then it will use the default
# (which may be random)
make_classical("rcbd")
#> 
#> ── experimental design details ─────────────────────────────────────────────────
#>   • This experimental design is often called Randomised Complete Block Design.
#>   • You can change the number in `set.seed` to get another random instance of
#>   the same design.
#>   • This design has a total of 99 units testing a total of 9 treatments.
#>   • This design is balanced and complete.
#> 
#> ── edibble code ────────────────────────────────────────────────────────────────
#> set.seed(762)
#> start_design("rcbd") %>%
#>   set_units(block = 11,
#>             unit = nested_in(block, 9)) %>%
#>   set_trts(treat = 9) %>%
#>   allot_trts(treat ~ unit) %>%
#>   assign_trts("random") %>%
#>   serve_table() 
#> 
#> ── edibble data frame ──────────────────────────────────────────────────────────
#> # An edibble: 99 x 3
#>         block       unit    treat
#>    <unit(11)> <unit(99)> <trt(9)>
#>  1     block1     unit1    treat7
#>  2     block1     unit2    treat8
#>  3     block1     unit3    treat5
#>  4     block1     unit4    treat2
#>  5     block1     unit5    treat4
#>  6     block1     unit6    treat9
#>  7     block1     unit7    treat1
#>  8     block1     unit8    treat3
#>  9     block1     unit9    treat6
#> 10     block2     unit10   treat7
#> # … with 89 more rows
# if you don't give any short names then it will generate a random one
make_classical()
#> → No name was supplied so selecting a random named experimental design...
#> Error in find_classical_names(): could not find function "find_classical_names"