Interactive plot of the edibble design

Interactive plot of the edibble design

Usage

# S3 method for edbl_design
plot(
  x,
  which = c("factors", "levels"),
  width = "100%",
  height = NULL,
  seed = 1,
  title = NULL,
  subtitle = NULL,
  footer = NULL,
  background = "transparent",
  view = c("show-buttons", "hide-buttons", "static"),
  ...
)

# S3 method for edbl_table
plot(x, ...)

plot_fct_graph(
  x,
  width = "100%",
  height = NULL,
  seed = 1,
  title = NULL,
  subtitle = NULL,
  footer = NULL,
  background = "transparent",
  view = c("show-buttons", "hide-buttons", "static"),
  ...
)

plot_lvl_graph(
  x,
  width = "100%",
  height = NULL,
  seed = 1,
  title = NULL,
  subtitle = NULL,
  footer = NULL,
  background = "transparent",
  view = c("show-buttons", "hide-buttons", "static"),
  ...
)

Arguments

x

An edibble design.

which

A string of either "factors" or "levels".

width, height

The width and height of the plot.

seed

A seed number so same plot is always generated.

title, subtitle, footer

The title, subtitle or footer of the plot. By default it uses the name from the x object as the title while rest is empty. To modify the look of the text, you can pass a character string consisting of valid for input style value in an HTML object, e.g. "font-size: 18px;font-family:serif;" as a named vector where the name corresponds to the text to display, e.g. c("Title" = "font-size:20px;").

background

The background color of the plot. Default is transparent. The input can be a color name (e.g. "white"), a HEX value ("#FFFFFF"), or rgb/rgba in the format like rgba(0, 0, 0, 0).

view

A string of either "show-buttons" (default), "hide-buttons", "static"

...

Currently unused.

Value

A plot.

Examples

plot(takeout(menu_crd(t = 4, n = 20)))
#> Loading required namespace: visNetwork