Print edibble graph to terminal

This function prints an edbl_graph object as a tree to terminal. The variables are color coded (or decorated) with the given options. Any ANSI coloring or styling are only visible in the console or terminal outputs that support it. The print output is best used interactively since any text styling are lost in text or R Markdown output. More details can be found in vignette("edbl-output").

# S3 method for edbl_graph
print(
  .data,
  decorate_units = edibble_decorate("units"),
  decorate_trts = edibble_decorate("trts"),
  decorate_resp = edibble_decorate("resp"),
  decorate_levels = edibble_decorate("levels"),
  decorate_main = edibble_decorate("main"),
  main = NULL
)

Arguments

.data
An edibble graph.

decorate_trts, decorate_units, decorate_resp, decorate_levels, decorate_title
A function applied to the name of treatment, unit, response factors or design title. The function should return a string. Most often this wraps the name with ANSI colored text. Run edibble_opt() to see the list of default values for the options.

Examples

# stylizing are only visible in terminal output that supports it
print(nclassics$split)

#> IGRAPH f9121e7 DN-- 20 51 -- Split plot design
#> + attr: name (g/c), vtype (v/c), vname (v/c), label (v/c), class (v/c),
#> | color (v/l), shape (v/c), label.color (v/c), label.family (v/c),
#> | label.font (v/n), label.cex (v/n), label2 (v/c), group (v/n), name
#> | (v/c), etype (e/c), color (e/c), arrow.mode (e/c), arrow.size (e/n),
#> | width (e/n), lty (e/n), curved (e/l)
#> + edges from f9121e7 (vertex names):
#> [1] subplot:subplot7->subplot:subplot8 subplot:subplot6->subplot:subplot7
#> [3] subplot:subplot5->subplot:subplot6 subplot:subplot4->subplot:subplot5
#> [5] subplot:subplot3->subplot:subplot4 subplot:subplot2->subplot:subplot3
#> [7] subplot:subplot1->subplot:subplot2 subplot         ->subplot:subplot8
#> + ... omitted several edges

## Split plot design
## ├─mainplot (4 levels)
## │ └─subplot (8 levels)
## ├─subplot (8 levels)
## ├─variety (2 levels)
## └─irrigation (2 levels)