Print edibble nexus to terminal

This function prints an edbl_nexus object as a tree to terminal. The variables are color coded (or decorated) with the given options. Any ANSI coloring or styling are only visible in the console or terminal outputs that support it. The print output is best used interactively since any text styling are lost in text or R Markdown output. More details can be found in vignette("edbl-output").

# S3 method for edbl_nexus
print(
  .nexus,
  decorate_units = edibble_decorate("units"),
  decorate_trts = edibble_decorate("trts"),
  decorate_resp = edibble_decorate("resp"),
  decorate_levels = edibble_decorate("levels"),
  decorate_main = edibble_decorate("main"),
  main = NULL
)

Arguments

.nexus
An edibble nexus.

decorate_trts, decorate_units, decorate_resp, decorate_levels, decorate_title
A function applied to the name of treatment, unit, response factors or design title. The function should return a string. Most often this wraps the name with ANSI colored text. Run edibble_opt() to see the list of default values for the options.

Examples

# stylizing are only visible in terminal output that supports it
print(nclassics$split)

#> Split plot design

## Split plot design
## ├─mainplot (4 levels)
## │ └─subplot (8 levels)
## ├─subplot (8 levels)
## ├─variety (2 levels)
## └─irrigation (2 levels)