Find the short names of the named designs

Find the short names of the named designs

Usage

scan_menu(packages = NULL, exclude = NULL)

Arguments

packages

A character vector containing the package names to search named designs from. By default it will search edibble and other packages loaded.

exclude

A character vector denoting the packages to exclude search from.

Value

A data.frame with package, name, arguments, and full name.

Examples

scan_menu()
#> # A tibble: 10 × 4
#>    package name         args                 name_full                          
#>    <chr>   <chr>        <chr>                <chr>                              
#>  1 edibble bibd         t, k, r, seed        Balanced Incomplete Block Design   
#>  2 edibble crd          t, n, r, seed        Completely Randomised Design       
#>  3 edibble factorial    trt, r, design, seed Factorial Design                   
#>  4 edibble graeco       t, seed              Graeco-Latin Square Design         
#>  5 edibble hyper_graeco t, seed              Hyper-Graeco-Latin Square Design   
#>  6 edibble lsd          t, seed              Latin Square Design                
#>  7 edibble rcbd         t, r, seed           Randomised Complete Block Design   
#>  8 edibble split        t1, t2, r, seed      Split-Plot Design, Split-Unit Desi…
#>  9 edibble strip        t1, t2, r, seed      Strip-Plot Design, Strip-Unit Desi…
#> 10 edibble youden       nc, t, seed          Youden Square Design