Split or count the data according to certain factors

This function has a similar result with split() where it returns a named list with names corresponding to the levels of the separating factor (or concatenated strings if multiple separating factors). The key differences to split(), are that the splitting factor does not appear in the elements of the list and only linked factors and their ancestors appear in the output, e.g. if treatment is applied to wholeplot and subplots are nested within subplots, then the subplot will not be shown in the output if split by treatment.

Usage

split_by(.data, ..., .sep = ":", .remove_empty = TRUE)

count_by(.data, ..., .remove_empty = TRUE)

Arguments

.data

An edibble table.

...

The factors to split or count by. You cannot split by a record factor or a factor that uniquely indexes the smallest unit in the design. You cannot also combine treatment and unit factors together.

.sep

The separator to use if more than one factor to split by.

.remove_empty

Remove empty combinations. Default is TRUE.

Value

A named list.

See also

Examples

spd <- takeout(menu_split())
spd %>% split(trt1)
#> Error: object 'trt1' not found
spd %>% split_by(trt1)
#> $trt11
#> # A tibble: 6 × 1
#>     mainplot
#>   <unit(18)>
#> 1 mainplot02
#> 2 mainplot06
#> 3 mainplot07
#> 4 mainplot08
#> 5 mainplot10
#> 6 mainplot18
#> 
#> $trt12
#> # A tibble: 6 × 1
#>     mainplot
#>   <unit(18)>
#> 1 mainplot01
#> 2 mainplot04
#> 3 mainplot09
#> 4 mainplot14
#> 5 mainplot15
#> 6 mainplot17
#> 
#> $trt13
#> # A tibble: 6 × 1
#>     mainplot
#>   <unit(18)>
#> 1 mainplot03
#> 2 mainplot05
#> 3 mainplot11
#> 4 mainplot12
#> 5 mainplot13
#> 6 mainplot16
#> 
spd %>% split_by(trt2)
#> $trt21
#> # A tibble: 18 × 2
#>       subplot   mainplot
#>    <unit(54)> <unit(18)>
#>  1  subplot01 mainplot01
#>  2  subplot05 mainplot02
#>  3  subplot07 mainplot03
#>  4  subplot12 mainplot04
#>  5  subplot15 mainplot05
#>  6  subplot16 mainplot06
#>  7  subplot19 mainplot07
#>  8  subplot23 mainplot08
#>  9  subplot25 mainplot09
#> 10  subplot30 mainplot10
#> 11  subplot31 mainplot11
#> 12  subplot34 mainplot12
#> 13  subplot38 mainplot13
#> 14  subplot41 mainplot14
#> 15  subplot43 mainplot15
#> 16  subplot48 mainplot16
#> 17  subplot50 mainplot17
#> 18  subplot52 mainplot18
#> 
#> $trt22
#> # A tibble: 18 × 2
#>       subplot   mainplot
#>    <unit(54)> <unit(18)>
#>  1  subplot03 mainplot01
#>  2  subplot06 mainplot02
#>  3  subplot09 mainplot03
#>  4  subplot10 mainplot04
#>  5  subplot14 mainplot05
#>  6  subplot18 mainplot06
#>  7  subplot20 mainplot07
#>  8  subplot24 mainplot08
#>  9  subplot26 mainplot09
#> 10  subplot28 mainplot10
#> 11  subplot32 mainplot11
#> 12  subplot36 mainplot12
#> 13  subplot39 mainplot13
#> 14  subplot40 mainplot14
#> 15  subplot45 mainplot15
#> 16  subplot47 mainplot16
#> 17  subplot51 mainplot17
#> 18  subplot54 mainplot18
#> 
#> $trt23
#> # A tibble: 18 × 2
#>       subplot   mainplot
#>    <unit(54)> <unit(18)>
#>  1  subplot02 mainplot01
#>  2  subplot04 mainplot02
#>  3  subplot08 mainplot03
#>  4  subplot11 mainplot04
#>  5  subplot13 mainplot05
#>  6  subplot17 mainplot06
#>  7  subplot21 mainplot07
#>  8  subplot22 mainplot08
#>  9  subplot27 mainplot09
#> 10  subplot29 mainplot10
#> 11  subplot33 mainplot11
#> 12  subplot35 mainplot12
#> 13  subplot37 mainplot13
#> 14  subplot42 mainplot14
#> 15  subplot44 mainplot15
#> 16  subplot46 mainplot16
#> 17  subplot49 mainplot17
#> 18  subplot53 mainplot18
#> 
spd %>% split_by(mainplot)
#> $mainplot01
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot01    trt21
#> 2  subplot02    trt23
#> 3  subplot03    trt22
#> 
#> $mainplot02
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot04    trt23
#> 2  subplot05    trt21
#> 3  subplot06    trt22
#> 
#> $mainplot03
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot07    trt21
#> 2  subplot08    trt23
#> 3  subplot09    trt22
#> 
#> $mainplot04
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot10    trt22
#> 2  subplot11    trt23
#> 3  subplot12    trt21
#> 
#> $mainplot05
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot13    trt23
#> 2  subplot14    trt22
#> 3  subplot15    trt21
#> 
#> $mainplot06
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot16    trt21
#> 2  subplot17    trt23
#> 3  subplot18    trt22
#> 
#> $mainplot07
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot19    trt21
#> 2  subplot20    trt22
#> 3  subplot21    trt23
#> 
#> $mainplot08
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot22    trt23
#> 2  subplot23    trt21
#> 3  subplot24    trt22
#> 
#> $mainplot09
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot25    trt21
#> 2  subplot26    trt22
#> 3  subplot27    trt23
#> 
#> $mainplot10
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot28    trt22
#> 2  subplot29    trt23
#> 3  subplot30    trt21
#> 
#> $mainplot11
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot31    trt21
#> 2  subplot32    trt22
#> 3  subplot33    trt23
#> 
#> $mainplot12
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot34    trt21
#> 2  subplot35    trt23
#> 3  subplot36    trt22
#> 
#> $mainplot13
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot37    trt23
#> 2  subplot38    trt21
#> 3  subplot39    trt22
#> 
#> $mainplot14
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot40    trt22
#> 2  subplot41    trt21
#> 3  subplot42    trt23
#> 
#> $mainplot15
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot43    trt21
#> 2  subplot44    trt23
#> 3  subplot45    trt22
#> 
#> $mainplot16
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot46    trt23
#> 2  subplot47    trt22
#> 3  subplot48    trt21
#> 
#> $mainplot17
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot49    trt23
#> 2  subplot50    trt21
#> 3  subplot51    trt22
#> 
#> $mainplot18
#> # A tibble: 3 × 2
#>      subplot     trt2
#>   <unit(54)> <trt(3)>
#> 1  subplot52    trt21
#> 2  subplot53    trt23
#> 3  subplot54    trt22
#> 
spd %>% count_by(trt1)
#>    trt1 mainplot
#> 1 trt11        6
#> 2 trt12        6
#> 3 trt13        6

fac <- takeout(menu_factorial(trt = c(2, 2, 2)))
fac %>% count_by(where(~is_trt(.x)))
#>    trt1  trt2  trt3 unit
#> 1 trt11 trt21 trt31    5
#> 2 trt12 trt21 trt31    5
#> 3 trt11 trt22 trt31    5
#> 4 trt12 trt22 trt31    5
#> 5 trt11 trt21 trt32    5
#> 6 trt12 trt21 trt32    5
#> 7 trt11 trt22 trt32    5
#> 8 trt12 trt22 trt32    5