Create a named experimental design

This function generates a named experimental design by supplying the selected menu named design and prints out by default

You can find the available recipes with scan_menu().

Usage

takeout(recipe = NULL, show = TRUE)

Arguments

recipe

A named design object. This should be typically generated from a function with prefix menu_. If nothing is supplied, it will randomly select one.

show

A logical value to indicate whether the code should be shown or not. Default is TRUE.

Value

A recipe design.

See also

Examples

takeout(menu_crd(n = 50, t = 5))
#> design("Completely Randomised Design") %>%
#>   set_units(unit = 50) %>%
#>   set_trts(trt = 5) %>%
#>   allot_trts(trt ~ unit) %>%
#>   assign_trts("random", seed = 81) %>%
#>   serve_table() 
#> 
#> # Completely Randomised Design 
#> # An edibble: 50 x 2
#>          unit      trt
#>  * <unit(50)> <trt(5)>
#>         <chr>    <chr>
#>  1     unit01     trt3
#>  2     unit02     trt3
#>  3     unit03     trt2
#>  4     unit04     trt5
#>  5     unit05     trt4
#>  6     unit06     trt3
#>  7     unit07     trt4
#>  8     unit08     trt4
#>  9     unit09     trt2
#> 10     unit10     trt4
#> # ℹ 40 more rows
# if you omit the design parameters then it will use the default
# (which may be random)
takeout(menu_crd())
#> design("Completely Randomised Design") %>%
#>   set_units(unit = 22) %>%
#>   set_trts(trt = 6) %>%
#>   allot_trts(trt ~ unit) %>%
#>   assign_trts("random", seed = 202) %>%
#>   serve_table() 
#> 
#> # Completely Randomised Design 
#> # An edibble: 22 x 2
#>          unit      trt
#>  * <unit(22)> <trt(6)>
#>         <chr>    <chr>
#>  1     unit01     trt5
#>  2     unit02     trt3
#>  3     unit03     trt1
#>  4     unit04     trt5
#>  5     unit05     trt1
#>  6     unit06     trt4
#>  7     unit07     trt2
#>  8     unit08     trt4
#>  9     unit09     trt6
#> 10     unit10     trt3
#> # ℹ 12 more rows
# if you don't give any short names then it will generate a random one
takeout()
#> → No name was supplied so selecting a random named experimental design...
#> → Selected Graeco-Latin Square Design
#> design("Graeco-Latin Square Design") %>%
#>   set_units(row = 4,
#>             col = 4,
#>             unit = crossed_by(row, col)) %>%
#>   set_trts(trt1 = 4,
#>            trt2 = 4) %>%
#>   allot_trts(trt1 ~ unit,
#>              trt2 ~ unit) %>%
#>   assign_trts("random", seed = 533) %>%
#>   serve_table() 
#> 
#> # Graeco-Latin Square Design 
#> # An edibble: 16 x 5
#>          row       col       unit     trt1     trt2
#>  * <unit(4)> <unit(4)> <unit(16)> <trt(4)> <trt(4)>
#>        <chr>     <chr>      <chr>    <chr>    <chr>
#>  1      row1      col1     unit01    trt12    trt21
#>  2      row2      col1     unit02    trt14    trt23
#>  3      row3      col1     unit03    trt11    trt24
#>  4      row4      col1     unit04    trt13    trt22
#>  5      row1      col2     unit05    trt13    trt22
#>  6      row2      col2     unit06    trt11    trt24
#>  7      row3      col2     unit07    trt14    trt21
#>  8      row4      col2     unit08    trt12    trt23
#>  9      row1      col3     unit09    trt11    trt24
#> 10      row2      col3     unit10    trt12    trt22
#> 11      row3      col3     unit11    trt13    trt23
#> 12      row4      col3     unit12    trt14    trt21
#> 13      row1      col4     unit13    trt14    trt23
#> 14      row2      col4     unit14    trt13    trt21
#> 15      row3      col4     unit15    trt12    trt22
#> 16      row4      col4     unit16    trt11    trt24