An R6 Class for an edibble design

An object with EdibbleDesign holds the core information related to the experimental design.

Active bindings

name

the name

graph

the name

table

the name

context

the name

muffle

the nam

active

Which is active

Methods

Public methods

EdibbleDesign$new()

EdibbleDesign$finalize()

EdibbleDesign$plot()

EdibbleDesign$print()

EdibbleDesign$save_seed()

EdibbleDesign$clone()

Method new()

Initialise the edibble design. See also start_design().

Usage

EdibbleDesign$new(name = NULL)

Arguments

name

The name or title of the project. The title should be informative and not overly long. If not supplied the default is "An edibble design".

Method finalize()

Invoked when the EdibbleDesign object is deleted.

Usage

EdibbleDesign$finalize()

Method plot()

Plot the intermediate contruct.

Usage

EdibbleDesign$plot(view = c("high", "low"), ..., main = NULL)

Arguments

view

A high- or -low level view.

...

Arguments passed into plot.igraph.

main

The title of the plot. By default it is the name of the EdibbleDesign object.

Method print()

Print the intermediate construct.

Usage

EdibbleDesign$print(...)

Arguments

...

Arguments passed into print.

Method save_seed()

This saves the .Random.seed used just prior to the randomisation process.

Usage

EdibbleDesign$save_seed()

Method clone()

The objects of this class are cloneable with this method.

Usage

EdibbleDesign$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.