edibble: Generate experimental designs using the grammar of experimental design

A system to facilitate statistical thinking of adapting experimental designs to different conditions.

Details

(WIP)

Website

The website for the package is at https://edibble.emitanaka.org

Discussion is at https://github.com/emitanaka/edibble/discussions

Package options

The following options are used for changing the default view for edibble graph or edibble design plot:

edibble.vertex.label.color.default

edibble.vertex.label.color.unit

The following options are used for changing the default view for the print out of edibble design or edibble graph.

edibble.tree.decorate.trts

edibble.tree.decorate.units

edibble.tree.decorate.rcrd

edibble.tree.decorate.levels

edibble.tree.decorate.main

TODO

See also

Author

Maintainer: Emi Tanaka dr.emi.tanaka@gmail.com (ORCID)