Create a classical named experimental design

The function make_classical generates a classical named experimental design by supplying its short name and prints out, by default:

.info: information about the named experimental design,

.code: code to create the design using edibble, and

.table: an edibble data frame for the generated design.

You can find the available short names with find_classical_names().

make_classical(
  .name = "",
  ...,
  .seed = as.integer(Sys.time()),
  .output = TRUE,
  .code = .output,
  .info = .output,
  .table = .output
)

code_classical(.name = "", .seed = as.integer(Sys.time()), ..., .quiet = FALSE)

Arguments

.name
The short name of the classical named experimental design. See under Details for the available named designs.

...
Parameters passed into the prep_classical_* functions.

.output
A short hand to turn all printing on or off. Default is TRUE.

.info, .code, .table
A logical to turn on or off for their respective parts.

.quiet
Opposite of .code. Whether to suppress code output.

Details

For The available named designs are:

"crd": completely randomised design

"rcbd": randomised complete block design

"split": split plot design

See also

Examples

make_classical("crd", n = 50, t = 5)

#> 

#> ── experimental design details ─────────────────────────────────────────────────

#> ● This experimental design is often called Completely Randomised Design.

#> ● You can change the number in `set.seed` to get another random instance of
#>   the same design.

#> ● This design has a total of 50 units testing a total of 5 treatments.

#> The following information is only true for the chosen parameters and not
#> necessary true for all Completely Randomised Designs.

#>   ◯ This design is balanced for the given numbers.

#> 

#>   ── edibble code ────────────────────────────────────────────────────────────────

#> set.seed(1610001077)
#> start_design("crd") %>%
#>   set_units(unit = 50) %>%
#>   set_trts(treat = 5) %>%
#>   allocate_trts(treat ~ unit) %>%
#>   randomise_trts() %>%
#>   serve_table() 

#> 

#>   ── edibble data frame ──────────────────────────────────────────────────────────

#> # An edibble: 50 x 2
#>    unit       treat   
#>    <unit(50)> <trt(5)>
#>  1 unit1      treat2  
#>  2 unit2      treat4  
#>  3 unit3      treat1  
#>  4 unit4      treat5  
#>  5 unit5      treat5  
#>  6 unit6      treat3  
#>  7 unit7      treat2  
#>  8 unit8      treat5  
#>  9 unit9      treat2  
#> 10 unit10     treat5  
#> # … with 40 more rows

# if you omit the design parameters then it will use the default
# (which may be random)
make_classical("rcbd")

#> 

#> ── experimental design details ─────────────────────────────────────────────────

#> ● This experimental design is often called Randomised Complete Block Design.

#> ● You can change the number in `set.seed` to get another random instance of
#>   the same design.

#> ● This design has a total of 66 units testing a total of 6 treatments.

#> ● This design is balanced and complete.

#> 

#> ── edibble code ────────────────────────────────────────────────────────────────

#> set.seed(1610001077)
#> start_design("rcbd") %>%
#>   set_units(block = 11,
#>             unit = nested_in(block, 6)) %>%
#>   set_trts(treat = 6) %>%
#>   allocate_trts(treat ~ unit) %>%
#>   randomise_trts() %>%
#>   serve_table() 

#> 

#> ── edibble data frame ──────────────────────────────────────────────────────────

#> # An edibble: 66 x 3
#>    block      unit       treat   
#>    <unit(11)> <unit(66)> <trt(6)>
#>  1 block1     unit1      treat5  
#>  2 block1     unit2      treat3  
#>  3 block1     unit3      treat3  
#>  4 block1     unit4      treat2  
#>  5 block1     unit5      treat1  
#>  6 block1     unit6      treat2  
#>  7 block2     unit7      treat1  
#>  8 block2     unit8      treat2  
#>  9 block2     unit9      treat5  
#> 10 block2     unit10     treat6  
#> # … with 56 more rows

# if you don't give any short names then it will generate a random one
make_classical()

#> → No name was supplied so selecting a random named experimental design...

#> 

#> ── experimental design details ─────────────────────────────────────────────────

#>   ◯ This experimental design is often called Completely Randomised Design.

#>   ◯ You can change the number in `set.seed` to get another random instance of
#>     the same design.

#>   ◯ This design has a total of 63 units testing a total of 3 treatments.

#>   The following information is only true for the chosen parameters and not
#>   necessary true for all Completely Randomised Designs.

#>   ─ This design is balanced for the given numbers.

#> 

#>   ── edibble code ────────────────────────────────────────────────────────────────

#> set.seed(1610001077)
#> start_design("crd") %>%
#>   set_units(unit = 63) %>%
#>   set_trts(treat = 3) %>%
#>   allocate_trts(treat ~ unit) %>%
#>   randomise_trts() %>%
#>   serve_table() 

#> 

#>   ── edibble data frame ──────────────────────────────────────────────────────────

#> # An edibble: 63 x 2
#>    unit       treat   
#>    <unit(63)> <trt(3)>
#>  1 unit1      treat1  
#>  2 unit2      treat2  
#>  3 unit3      treat2  
#>  4 unit4      treat2  
#>  5 unit5      treat2  
#>  6 unit6      treat3  
#>  7 unit7      treat3  
#>  8 unit8      treat1  
#>  9 unit9      treat1  
#> 10 unit10     treat2  
#> # … with 53 more rows