Create a classical named experimental design

This function is defunct. Please use takeout instead.

The function make_classical generates a classical named experimental design by supplying its short name and prints out, by default:

info: information about the named experimental design,

code: code to create the design using edibble, and

table: an edibble data frame for the generated design.

You can find the available short names with find_classical_names().

Usage

make_classical(.name = "", ..., .output = TRUE)

code_classical(.name = "", ..., .quiet = FALSE)

Arguments

.name

The short name of the classical named experimental design. See under Details for the available named designs.

...

Parameters passed into the prep_classical_* functions.

.output

A logical value to indicate whether all output should be printed or not or a vector of character (e.g. c("info", "code", "table")) specifying which of the three outputs should be printed. Default is TRUE.

Details

For The available named designs are:

"crd": completely randomised design

"rcbd": randomised complete block design

"split": split plot design

See also

Examples

make_classical("crd", n = 50, t = 5)
#> Warning: `prep_classical_crd` is deprecated. Please use `menu_crd` instead.
#> 
#> ── experimental design details ──
#> 
#>   • This experimental design is often called Completely Randomised Design.
#>   • You can change the number in `seed` to get another random instance of the
#>   same design.
#> 
#> ── edibble code ──
#> 
#> start_design("Completely Randomised Design") %>%
#>   set_units(unit = 50) %>%
#>   set_trts(trt = 5) %>%
#>   allot_trts(trt ~ unit) %>%
#>   assign_trts("random", seed = 334) %>%
#>   serve_table() 
#> ── edibble data frame ──
#> 
#> # Completely Randomised Design 
#> # An edibble: 50 x 2
#>          unit      trt
#>    <unit(50)> <trt(5)>
#>  1     unit1      trt2
#>  2     unit2      trt4
#>  3     unit3      trt3
#>  4     unit4      trt2
#>  5     unit5      trt5
#>  6     unit6      trt5
#>  7     unit7      trt4
#>  8     unit8      trt1
#>  9     unit9      trt2
#> 10     unit10     trt3
#> # … with 40 more rows
# if you omit the design parameters then it will use the default
# (which may be random)
make_classical("rcbd")
#> Warning: `prep_classical_rcbd` is deprecated. Please use `menu_rcbd` instead.
#> ── experimental design details ──
#> 
#>   • This experimental design is often called Randomised Complete Block Design.
#>   • You can change the number in `seed` to get another random instance of the
#>   same design.
#> 
#> ── edibble code ──
#> 
#> start_design("Randomised Complete Block Design") %>%
#>   set_units(block = 5,
#>             unit = nested_in(block, 2)) %>%
#>   set_trts(trt = 2) %>%
#>   allot_trts(trt ~ unit) %>%
#>   assign_trts("random", seed = 105) %>%
#>   serve_table() 
#> ── edibble data frame ──
#> 
#> # Randomised Complete Block Design 
#> # An edibble: 10 x 3
#>        block       unit      trt
#>    <unit(5)> <unit(10)> <trt(2)>
#>  1    block1     unit1      trt1
#>  2    block1     unit2      trt2
#>  3    block2     unit3      trt1
#>  4    block2     unit4      trt2
#>  5    block3     unit5      trt1
#>  6    block3     unit6      trt2
#>  7    block4     unit7      trt2
#>  8    block4     unit8      trt1
#>  9    block5     unit9      trt2
#> 10    block5     unit10     trt1
# if you don't give any short names then it will generate a random one
make_classical()
#> → No name was supplied so selecting a random named experimental design...
#> Warning: This function is defunct. Use `scan_menu` instead.
#> Warning: `prep_classical_crd` is deprecated. Please use `menu_crd` instead.
#> Warning: `prep_classical_factorial` is deprecated. Please use `menu_factorial` instead.
#> Error in match.arg(design): 'arg' must be NULL or a character vector