Split or count the data according to certain factors

This function has a similar result with split() where it returns a named list with names corresponding to the levels of the separating factor (or concatenated strings if multiple separating factors). The key differences to split(), are that the splitting factor does not appear in the elements of the list and only linked factors and their ancestors appear in the output, e.g. if treatment is applied to wholeplot and subplots are nested within subplots, then the subplot will not be shown in the output if split by treatment.

Usage

split_by(.data, ..., .sep = ":", .remove_empty = TRUE)

count_by(.data, ..., .remove_empty = TRUE)

Arguments

.data

An edibble table.

...

The factors to split or count by. You cannot split by a record factor or a factor that uniquely indexes the smallest unit in the design. You cannot also combine treatment and unit factors together.

.sep

The separator to use if more than one factor to split by.

.remove_empty

Remove empty combinations. Default is TRUE.

Value

A named list.

See also

Examples

spd <- takeout(menu_split())
split(spd, spd$trt1)
#> $trt11
#> design("Split-Plot Design | Split-Unit Design") %>%
#>   set_units(mainplot = 90,
#>              subplot = nested_in(mainplot, 5)) %>%
#>   set_trts(trt1 = 9,
#>            trt2 = 5) %>%
#>   allot_trts(trt1 ~ mainplot,
#>              trt2 ~ subplot) %>%
#>   assign_trts("random", seed = 546) %>%
#>   serve_table() 
#> 
#> # Split-Plot Design | Split-Unit Design 
#> # An edibble: 50 x 4
#>      mainplot    subplot   trt1   trt2
#>       <U(90)>   <U(450)> <T(9)> <T(5)>
#>         <chr>      <chr>  <chr>  <chr>
#>  1 mainplot10 subplot046  trt11  trt21
#>  2 mainplot10 subplot047  trt11  trt22
#>  3 mainplot10 subplot048  trt11  trt24
#>  4 mainplot10 subplot049  trt11  trt25
#>  5 mainplot10 subplot050  trt11  trt23
#>  6 mainplot13 subplot061  trt11  trt22
#>  7 mainplot13 subplot062  trt11  trt25
#>  8 mainplot13 subplot063  trt11  trt23
#>  9 mainplot13 subplot064  trt11  trt21
#> 10 mainplot13 subplot065  trt11  trt24
#> # ℹ 40 more rows
#> 
#> $trt12
#> design("Split-Plot Design | Split-Unit Design") %>%
#>   set_units(mainplot = 90,
#>              subplot = nested_in(mainplot, 5)) %>%
#>   set_trts(trt1 = 9,
#>            trt2 = 5) %>%
#>   allot_trts(trt1 ~ mainplot,
#>              trt2 ~ subplot) %>%
#>   assign_trts("random", seed = 546) %>%
#>   serve_table() 
#> 
#> # Split-Plot Design | Split-Unit Design 
#> # An edibble: 50 x 4
#>      mainplot    subplot   trt1   trt2
#>       <U(90)>   <U(450)> <T(9)> <T(5)>
#>         <chr>      <chr>  <chr>  <chr>
#>  1 mainplot02 subplot006  trt12  trt22
#>  2 mainplot02 subplot007  trt12  trt24
#>  3 mainplot02 subplot008  trt12  trt21
#>  4 mainplot02 subplot009  trt12  trt23
#>  5 mainplot02 subplot010  trt12  trt25
#>  6 mainplot04 subplot016  trt12  trt22
#>  7 mainplot04 subplot017  trt12  trt23
#>  8 mainplot04 subplot018  trt12  trt24
#>  9 mainplot04 subplot019  trt12  trt25
#> 10 mainplot04 subplot020  trt12  trt21
#> # ℹ 40 more rows
#> 
#> $trt13
#> design("Split-Plot Design | Split-Unit Design") %>%
#>   set_units(mainplot = 90,
#>              subplot = nested_in(mainplot, 5)) %>%
#>   set_trts(trt1 = 9,
#>            trt2 = 5) %>%
#>   allot_trts(trt1 ~ mainplot,
#>              trt2 ~ subplot) %>%
#>   assign_trts("random", seed = 546) %>%
#>   serve_table() 
#> 
#> # Split-Plot Design | Split-Unit Design 
#> # An edibble: 50 x 4
#>      mainplot    subplot   trt1   trt2
#>       <U(90)>   <U(450)> <T(9)> <T(5)>
#>         <chr>      <chr>  <chr>  <chr>
#>  1 mainplot14 subplot066  trt13  trt24
#>  2 mainplot14 subplot067  trt13  trt22
#>  3 mainplot14 subplot068  trt13  trt23
#>  4 mainplot14 subplot069  trt13  trt21
#>  5 mainplot14 subplot070  trt13  trt25
#>  6 mainplot40 subplot196  trt13  trt22
#>  7 mainplot40 subplot197  trt13  trt23
#>  8 mainplot40 subplot198  trt13  trt25
#>  9 mainplot40 subplot199  trt13  trt21
#> 10 mainplot40 subplot200  trt13  trt24
#> # ℹ 40 more rows
#> 
#> $trt14
#> design("Split-Plot Design | Split-Unit Design") %>%
#>   set_units(mainplot = 90,
#>              subplot = nested_in(mainplot, 5)) %>%
#>   set_trts(trt1 = 9,
#>            trt2 = 5) %>%
#>   allot_trts(trt1 ~ mainplot,
#>              trt2 ~ subplot) %>%
#>   assign_trts("random", seed = 546) %>%
#>   serve_table() 
#> 
#> # Split-Plot Design | Split-Unit Design 
#> # An edibble: 50 x 4
#>      mainplot    subplot   trt1   trt2
#>       <U(90)>   <U(450)> <T(9)> <T(5)>
#>         <chr>      <chr>  <chr>  <chr>
#>  1 mainplot06 subplot026  trt14  trt22
#>  2 mainplot06 subplot027  trt14  trt25
#>  3 mainplot06 subplot028  trt14  trt24
#>  4 mainplot06 subplot029  trt14  trt21
#>  5 mainplot06 subplot030  trt14  trt23
#>  6 mainplot09 subplot041  trt14  trt25
#>  7 mainplot09 subplot042  trt14  trt23
#>  8 mainplot09 subplot043  trt14  trt21
#>  9 mainplot09 subplot044  trt14  trt22
#> 10 mainplot09 subplot045  trt14  trt24
#> # ℹ 40 more rows
#> 
#> $trt15
#> design("Split-Plot Design | Split-Unit Design") %>%
#>   set_units(mainplot = 90,
#>              subplot = nested_in(mainplot, 5)) %>%
#>   set_trts(trt1 = 9,
#>            trt2 = 5) %>%
#>   allot_trts(trt1 ~ mainplot,
#>              trt2 ~ subplot) %>%
#>   assign_trts("random", seed = 546) %>%
#>   serve_table() 
#> 
#> # Split-Plot Design | Split-Unit Design 
#> # An edibble: 50 x 4
#>      mainplot    subplot   trt1   trt2
#>       <U(90)>   <U(450)> <T(9)> <T(5)>
#>         <chr>      <chr>  <chr>  <chr>
#>  1 mainplot01 subplot001  trt15  trt24
#>  2 mainplot01 subplot002  trt15  trt22
#>  3 mainplot01 subplot003  trt15  trt25
#>  4 mainplot01 subplot004  trt15  trt21
#>  5 mainplot01 subplot005  trt15  trt23
#>  6 mainplot18 subplot086  trt15  trt24
#>  7 mainplot18 subplot087  trt15  trt21
#>  8 mainplot18 subplot088  trt15  trt22
#>  9 mainplot18 subplot089  trt15  trt25
#> 10 mainplot18 subplot090  trt15  trt23
#> # ℹ 40 more rows
#> 
#> $trt16
#> design("Split-Plot Design | Split-Unit Design") %>%
#>   set_units(mainplot = 90,
#>              subplot = nested_in(mainplot, 5)) %>%
#>   set_trts(trt1 = 9,
#>            trt2 = 5) %>%
#>   allot_trts(trt1 ~ mainplot,
#>              trt2 ~ subplot) %>%
#>   assign_trts("random", seed = 546) %>%
#>   serve_table() 
#> 
#> # Split-Plot Design | Split-Unit Design 
#> # An edibble: 50 x 4
#>      mainplot    subplot   trt1   trt2
#>       <U(90)>   <U(450)> <T(9)> <T(5)>
#>         <chr>      <chr>  <chr>  <chr>
#>  1 mainplot15 subplot071  trt16  trt25
#>  2 mainplot15 subplot072  trt16  trt21
#>  3 mainplot15 subplot073  trt16  trt24
#>  4 mainplot15 subplot074  trt16  trt23
#>  5 mainplot15 subplot075  trt16  trt22
#>  6 mainplot16 subplot076  trt16  trt25
#>  7 mainplot16 subplot077  trt16  trt22
#>  8 mainplot16 subplot078  trt16  trt21
#>  9 mainplot16 subplot079  trt16  trt23
#> 10 mainplot16 subplot080  trt16  trt24
#> # ℹ 40 more rows
#> 
#> $trt17
#> design("Split-Plot Design | Split-Unit Design") %>%
#>   set_units(mainplot = 90,
#>              subplot = nested_in(mainplot, 5)) %>%
#>   set_trts(trt1 = 9,
#>            trt2 = 5) %>%
#>   allot_trts(trt1 ~ mainplot,
#>              trt2 ~ subplot) %>%
#>   assign_trts("random", seed = 546) %>%
#>   serve_table() 
#> 
#> # Split-Plot Design | Split-Unit Design 
#> # An edibble: 50 x 4
#>      mainplot    subplot   trt1   trt2
#>       <U(90)>   <U(450)> <T(9)> <T(5)>
#>         <chr>      <chr>  <chr>  <chr>
#>  1 mainplot17 subplot081  trt17  trt25
#>  2 mainplot17 subplot082  trt17  trt22
#>  3 mainplot17 subplot083  trt17  trt23
#>  4 mainplot17 subplot084  trt17  trt24
#>  5 mainplot17 subplot085  trt17  trt21
#>  6 mainplot36 subplot176  trt17  trt25
#>  7 mainplot36 subplot177  trt17  trt22
#>  8 mainplot36 subplot178  trt17  trt24
#>  9 mainplot36 subplot179  trt17  trt21
#> 10 mainplot36 subplot180  trt17  trt23
#> # ℹ 40 more rows
#> 
#> $trt18
#> design("Split-Plot Design | Split-Unit Design") %>%
#>   set_units(mainplot = 90,
#>              subplot = nested_in(mainplot, 5)) %>%
#>   set_trts(trt1 = 9,
#>            trt2 = 5) %>%
#>   allot_trts(trt1 ~ mainplot,
#>              trt2 ~ subplot) %>%
#>   assign_trts("random", seed = 546) %>%
#>   serve_table() 
#> 
#> # Split-Plot Design | Split-Unit Design 
#> # An edibble: 50 x 4
#>      mainplot    subplot   trt1   trt2
#>       <U(90)>   <U(450)> <T(9)> <T(5)>
#>         <chr>      <chr>  <chr>  <chr>
#>  1 mainplot07 subplot031  trt18  trt21
#>  2 mainplot07 subplot032  trt18  trt23
#>  3 mainplot07 subplot033  trt18  trt22
#>  4 mainplot07 subplot034  trt18  trt24
#>  5 mainplot07 subplot035  trt18  trt25
#>  6 mainplot11 subplot051  trt18  trt22
#>  7 mainplot11 subplot052  trt18  trt25
#>  8 mainplot11 subplot053  trt18  trt21
#>  9 mainplot11 subplot054  trt18  trt23
#> 10 mainplot11 subplot055  trt18  trt24
#> # ℹ 40 more rows
#> 
#> $trt19
#> design("Split-Plot Design | Split-Unit Design") %>%
#>   set_units(mainplot = 90,
#>              subplot = nested_in(mainplot, 5)) %>%
#>   set_trts(trt1 = 9,
#>            trt2 = 5) %>%
#>   allot_trts(trt1 ~ mainplot,
#>              trt2 ~ subplot) %>%
#>   assign_trts("random", seed = 546) %>%
#>   serve_table() 
#> 
#> # Split-Plot Design | Split-Unit Design 
#> # An edibble: 50 x 4
#>      mainplot    subplot   trt1   trt2
#>       <U(90)>   <U(450)> <T(9)> <T(5)>
#>         <chr>      <chr>  <chr>  <chr>
#>  1 mainplot03 subplot011  trt19  trt22
#>  2 mainplot03 subplot012  trt19  trt24
#>  3 mainplot03 subplot013  trt19  trt23
#>  4 mainplot03 subplot014  trt19  trt21
#>  5 mainplot03 subplot015  trt19  trt25
#>  6 mainplot08 subplot036  trt19  trt24
#>  7 mainplot08 subplot037  trt19  trt21
#>  8 mainplot08 subplot038  trt19  trt25
#>  9 mainplot08 subplot039  trt19  trt23
#> 10 mainplot08 subplot040  trt19  trt22
#> # ℹ 40 more rows
#> 
spd %>% split_by(trt1)
#> $trt11
#> # A tibble: 10 × 1
#>      mainplot
#>       <U(90)>
#>  1 mainplot10
#>  2 mainplot13
#>  3 mainplot38
#>  4 mainplot43
#>  5 mainplot50
#>  6 mainplot70
#>  7 mainplot79
#>  8 mainplot81
#>  9 mainplot88
#> 10 mainplot90
#> 
#> $trt12
#> # A tibble: 10 × 1
#>      mainplot
#>       <U(90)>
#>  1 mainplot02
#>  2 mainplot04
#>  3 mainplot05
#>  4 mainplot22
#>  5 mainplot32
#>  6 mainplot42
#>  7 mainplot65
#>  8 mainplot73
#>  9 mainplot74
#> 10 mainplot77
#> 
#> $trt13
#> # A tibble: 10 × 1
#>      mainplot
#>       <U(90)>
#>  1 mainplot14
#>  2 mainplot40
#>  3 mainplot46
#>  4 mainplot54
#>  5 mainplot56
#>  6 mainplot57
#>  7 mainplot58
#>  8 mainplot63
#>  9 mainplot64
#> 10 mainplot76
#> 
#> $trt14
#> # A tibble: 10 × 1
#>      mainplot
#>       <U(90)>
#>  1 mainplot06
#>  2 mainplot09
#>  3 mainplot20
#>  4 mainplot24
#>  5 mainplot27
#>  6 mainplot31
#>  7 mainplot51
#>  8 mainplot61
#>  9 mainplot72
#> 10 mainplot86
#> 
#> $trt15
#> # A tibble: 10 × 1
#>      mainplot
#>       <U(90)>
#>  1 mainplot01
#>  2 mainplot18
#>  3 mainplot19
#>  4 mainplot26
#>  5 mainplot34
#>  6 mainplot47
#>  7 mainplot48
#>  8 mainplot66
#>  9 mainplot71
#> 10 mainplot83
#> 
#> $trt16
#> # A tibble: 10 × 1
#>      mainplot
#>       <U(90)>
#>  1 mainplot15
#>  2 mainplot16
#>  3 mainplot30
#>  4 mainplot41
#>  5 mainplot45
#>  6 mainplot59
#>  7 mainplot60
#>  8 mainplot62
#>  9 mainplot80
#> 10 mainplot89
#> 
#> $trt17
#> # A tibble: 10 × 1
#>      mainplot
#>       <U(90)>
#>  1 mainplot17
#>  2 mainplot36
#>  3 mainplot37
#>  4 mainplot52
#>  5 mainplot53
#>  6 mainplot67
#>  7 mainplot68
#>  8 mainplot75
#>  9 mainplot78
#> 10 mainplot85
#> 
#> $trt18
#> # A tibble: 10 × 1
#>      mainplot
#>       <U(90)>
#>  1 mainplot07
#>  2 mainplot11
#>  3 mainplot21
#>  4 mainplot23
#>  5 mainplot29
#>  6 mainplot39
#>  7 mainplot55
#>  8 mainplot69
#>  9 mainplot84
#> 10 mainplot87
#> 
#> $trt19
#> # A tibble: 10 × 1
#>      mainplot
#>       <U(90)>
#>  1 mainplot03
#>  2 mainplot08
#>  3 mainplot12
#>  4 mainplot25
#>  5 mainplot28
#>  6 mainplot33
#>  7 mainplot35
#>  8 mainplot44
#>  9 mainplot49
#> 10 mainplot82
#> 
spd %>% split_by(trt2)
#> $trt21
#> # A tibble: 90 × 2
#>       subplot   mainplot
#>      <U(450)>    <U(90)>
#>  1 subplot004 mainplot01
#>  2 subplot008 mainplot02
#>  3 subplot014 mainplot03
#>  4 subplot020 mainplot04
#>  5 subplot022 mainplot05
#>  6 subplot029 mainplot06
#>  7 subplot031 mainplot07
#>  8 subplot037 mainplot08
#>  9 subplot043 mainplot09
#> 10 subplot046 mainplot10
#> # ℹ 80 more rows
#> 
#> $trt22
#> # A tibble: 90 × 2
#>       subplot   mainplot
#>      <U(450)>    <U(90)>
#>  1 subplot002 mainplot01
#>  2 subplot006 mainplot02
#>  3 subplot011 mainplot03
#>  4 subplot016 mainplot04
#>  5 subplot024 mainplot05
#>  6 subplot026 mainplot06
#>  7 subplot033 mainplot07
#>  8 subplot040 mainplot08
#>  9 subplot044 mainplot09
#> 10 subplot047 mainplot10
#> # ℹ 80 more rows
#> 
#> $trt23
#> # A tibble: 90 × 2
#>       subplot   mainplot
#>      <U(450)>    <U(90)>
#>  1 subplot005 mainplot01
#>  2 subplot009 mainplot02
#>  3 subplot013 mainplot03
#>  4 subplot017 mainplot04
#>  5 subplot023 mainplot05
#>  6 subplot030 mainplot06
#>  7 subplot032 mainplot07
#>  8 subplot039 mainplot08
#>  9 subplot042 mainplot09
#> 10 subplot050 mainplot10
#> # ℹ 80 more rows
#> 
#> $trt24
#> # A tibble: 90 × 2
#>       subplot   mainplot
#>      <U(450)>    <U(90)>
#>  1 subplot001 mainplot01
#>  2 subplot007 mainplot02
#>  3 subplot012 mainplot03
#>  4 subplot018 mainplot04
#>  5 subplot025 mainplot05
#>  6 subplot028 mainplot06
#>  7 subplot034 mainplot07
#>  8 subplot036 mainplot08
#>  9 subplot045 mainplot09
#> 10 subplot048 mainplot10
#> # ℹ 80 more rows
#> 
#> $trt25
#> # A tibble: 90 × 2
#>       subplot   mainplot
#>      <U(450)>    <U(90)>
#>  1 subplot003 mainplot01
#>  2 subplot010 mainplot02
#>  3 subplot015 mainplot03
#>  4 subplot019 mainplot04
#>  5 subplot021 mainplot05
#>  6 subplot027 mainplot06
#>  7 subplot035 mainplot07
#>  8 subplot038 mainplot08
#>  9 subplot041 mainplot09
#> 10 subplot049 mainplot10
#> # ℹ 80 more rows
#> 
spd %>% split_by(mainplot)
#> $mainplot01
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot001  trt24
#> 2 subplot002  trt22
#> 3 subplot003  trt25
#> 4 subplot004  trt21
#> 5 subplot005  trt23
#> 
#> $mainplot02
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot006  trt22
#> 2 subplot007  trt24
#> 3 subplot008  trt21
#> 4 subplot009  trt23
#> 5 subplot010  trt25
#> 
#> $mainplot03
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot011  trt22
#> 2 subplot012  trt24
#> 3 subplot013  trt23
#> 4 subplot014  trt21
#> 5 subplot015  trt25
#> 
#> $mainplot04
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot016  trt22
#> 2 subplot017  trt23
#> 3 subplot018  trt24
#> 4 subplot019  trt25
#> 5 subplot020  trt21
#> 
#> $mainplot05
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot021  trt25
#> 2 subplot022  trt21
#> 3 subplot023  trt23
#> 4 subplot024  trt22
#> 5 subplot025  trt24
#> 
#> $mainplot06
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot026  trt22
#> 2 subplot027  trt25
#> 3 subplot028  trt24
#> 4 subplot029  trt21
#> 5 subplot030  trt23
#> 
#> $mainplot07
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot031  trt21
#> 2 subplot032  trt23
#> 3 subplot033  trt22
#> 4 subplot034  trt24
#> 5 subplot035  trt25
#> 
#> $mainplot08
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot036  trt24
#> 2 subplot037  trt21
#> 3 subplot038  trt25
#> 4 subplot039  trt23
#> 5 subplot040  trt22
#> 
#> $mainplot09
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot041  trt25
#> 2 subplot042  trt23
#> 3 subplot043  trt21
#> 4 subplot044  trt22
#> 5 subplot045  trt24
#> 
#> $mainplot10
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot046  trt21
#> 2 subplot047  trt22
#> 3 subplot048  trt24
#> 4 subplot049  trt25
#> 5 subplot050  trt23
#> 
#> $mainplot11
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot051  trt22
#> 2 subplot052  trt25
#> 3 subplot053  trt21
#> 4 subplot054  trt23
#> 5 subplot055  trt24
#> 
#> $mainplot12
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot056  trt23
#> 2 subplot057  trt25
#> 3 subplot058  trt22
#> 4 subplot059  trt21
#> 5 subplot060  trt24
#> 
#> $mainplot13
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot061  trt22
#> 2 subplot062  trt25
#> 3 subplot063  trt23
#> 4 subplot064  trt21
#> 5 subplot065  trt24
#> 
#> $mainplot14
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot066  trt24
#> 2 subplot067  trt22
#> 3 subplot068  trt23
#> 4 subplot069  trt21
#> 5 subplot070  trt25
#> 
#> $mainplot15
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot071  trt25
#> 2 subplot072  trt21
#> 3 subplot073  trt24
#> 4 subplot074  trt23
#> 5 subplot075  trt22
#> 
#> $mainplot16
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot076  trt25
#> 2 subplot077  trt22
#> 3 subplot078  trt21
#> 4 subplot079  trt23
#> 5 subplot080  trt24
#> 
#> $mainplot17
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot081  trt25
#> 2 subplot082  trt22
#> 3 subplot083  trt23
#> 4 subplot084  trt24
#> 5 subplot085  trt21
#> 
#> $mainplot18
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot086  trt24
#> 2 subplot087  trt21
#> 3 subplot088  trt22
#> 4 subplot089  trt25
#> 5 subplot090  trt23
#> 
#> $mainplot19
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot091  trt21
#> 2 subplot092  trt24
#> 3 subplot093  trt25
#> 4 subplot094  trt23
#> 5 subplot095  trt22
#> 
#> $mainplot20
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot096  trt25
#> 2 subplot097  trt24
#> 3 subplot098  trt23
#> 4 subplot099  trt21
#> 5 subplot100  trt22
#> 
#> $mainplot21
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot101  trt23
#> 2 subplot102  trt25
#> 3 subplot103  trt24
#> 4 subplot104  trt22
#> 5 subplot105  trt21
#> 
#> $mainplot22
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot106  trt21
#> 2 subplot107  trt24
#> 3 subplot108  trt22
#> 4 subplot109  trt23
#> 5 subplot110  trt25
#> 
#> $mainplot23
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot111  trt25
#> 2 subplot112  trt24
#> 3 subplot113  trt22
#> 4 subplot114  trt23
#> 5 subplot115  trt21
#> 
#> $mainplot24
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot116  trt21
#> 2 subplot117  trt23
#> 3 subplot118  trt22
#> 4 subplot119  trt24
#> 5 subplot120  trt25
#> 
#> $mainplot25
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot121  trt24
#> 2 subplot122  trt22
#> 3 subplot123  trt23
#> 4 subplot124  trt21
#> 5 subplot125  trt25
#> 
#> $mainplot26
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot126  trt24
#> 2 subplot127  trt22
#> 3 subplot128  trt23
#> 4 subplot129  trt21
#> 5 subplot130  trt25
#> 
#> $mainplot27
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot131  trt23
#> 2 subplot132  trt25
#> 3 subplot133  trt24
#> 4 subplot134  trt21
#> 5 subplot135  trt22
#> 
#> $mainplot28
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot136  trt23
#> 2 subplot137  trt21
#> 3 subplot138  trt24
#> 4 subplot139  trt22
#> 5 subplot140  trt25
#> 
#> $mainplot29
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot141  trt24
#> 2 subplot142  trt22
#> 3 subplot143  trt23
#> 4 subplot144  trt21
#> 5 subplot145  trt25
#> 
#> $mainplot30
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot146  trt25
#> 2 subplot147  trt23
#> 3 subplot148  trt24
#> 4 subplot149  trt21
#> 5 subplot150  trt22
#> 
#> $mainplot31
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot151  trt24
#> 2 subplot152  trt22
#> 3 subplot153  trt23
#> 4 subplot154  trt21
#> 5 subplot155  trt25
#> 
#> $mainplot32
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot156  trt22
#> 2 subplot157  trt23
#> 3 subplot158  trt25
#> 4 subplot159  trt24
#> 5 subplot160  trt21
#> 
#> $mainplot33
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot161  trt21
#> 2 subplot162  trt24
#> 3 subplot163  trt23
#> 4 subplot164  trt25
#> 5 subplot165  trt22
#> 
#> $mainplot34
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot166  trt25
#> 2 subplot167  trt24
#> 3 subplot168  trt22
#> 4 subplot169  trt23
#> 5 subplot170  trt21
#> 
#> $mainplot35
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot171  trt21
#> 2 subplot172  trt25
#> 3 subplot173  trt24
#> 4 subplot174  trt22
#> 5 subplot175  trt23
#> 
#> $mainplot36
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot176  trt25
#> 2 subplot177  trt22
#> 3 subplot178  trt24
#> 4 subplot179  trt21
#> 5 subplot180  trt23
#> 
#> $mainplot37
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot181  trt24
#> 2 subplot182  trt22
#> 3 subplot183  trt23
#> 4 subplot184  trt21
#> 5 subplot185  trt25
#> 
#> $mainplot38
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot186  trt23
#> 2 subplot187  trt24
#> 3 subplot188  trt21
#> 4 subplot189  trt25
#> 5 subplot190  trt22
#> 
#> $mainplot39
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot191  trt24
#> 2 subplot192  trt21
#> 3 subplot193  trt25
#> 4 subplot194  trt22
#> 5 subplot195  trt23
#> 
#> $mainplot40
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot196  trt22
#> 2 subplot197  trt23
#> 3 subplot198  trt25
#> 4 subplot199  trt21
#> 5 subplot200  trt24
#> 
#> $mainplot41
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot201  trt24
#> 2 subplot202  trt21
#> 3 subplot203  trt25
#> 4 subplot204  trt23
#> 5 subplot205  trt22
#> 
#> $mainplot42
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot206  trt21
#> 2 subplot207  trt23
#> 3 subplot208  trt22
#> 4 subplot209  trt24
#> 5 subplot210  trt25
#> 
#> $mainplot43
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot211  trt23
#> 2 subplot212  trt24
#> 3 subplot213  trt21
#> 4 subplot214  trt25
#> 5 subplot215  trt22
#> 
#> $mainplot44
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot216  trt23
#> 2 subplot217  trt25
#> 3 subplot218  trt24
#> 4 subplot219  trt21
#> 5 subplot220  trt22
#> 
#> $mainplot45
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot221  trt24
#> 2 subplot222  trt21
#> 3 subplot223  trt23
#> 4 subplot224  trt22
#> 5 subplot225  trt25
#> 
#> $mainplot46
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot226  trt23
#> 2 subplot227  trt24
#> 3 subplot228  trt22
#> 4 subplot229  trt25
#> 5 subplot230  trt21
#> 
#> $mainplot47
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot231  trt25
#> 2 subplot232  trt21
#> 3 subplot233  trt23
#> 4 subplot234  trt22
#> 5 subplot235  trt24
#> 
#> $mainplot48
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot236  trt24
#> 2 subplot237  trt23
#> 3 subplot238  trt25
#> 4 subplot239  trt21
#> 5 subplot240  trt22
#> 
#> $mainplot49
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot241  trt24
#> 2 subplot242  trt21
#> 3 subplot243  trt25
#> 4 subplot244  trt22
#> 5 subplot245  trt23
#> 
#> $mainplot50
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot246  trt25
#> 2 subplot247  trt21
#> 3 subplot248  trt22
#> 4 subplot249  trt24
#> 5 subplot250  trt23
#> 
#> $mainplot51
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot251  trt24
#> 2 subplot252  trt22
#> 3 subplot253  trt23
#> 4 subplot254  trt25
#> 5 subplot255  trt21
#> 
#> $mainplot52
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot256  trt24
#> 2 subplot257  trt22
#> 3 subplot258  trt21
#> 4 subplot259  trt23
#> 5 subplot260  trt25
#> 
#> $mainplot53
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot261  trt21
#> 2 subplot262  trt23
#> 3 subplot263  trt24
#> 4 subplot264  trt25
#> 5 subplot265  trt22
#> 
#> $mainplot54
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot266  trt23
#> 2 subplot267  trt24
#> 3 subplot268  trt25
#> 4 subplot269  trt22
#> 5 subplot270  trt21
#> 
#> $mainplot55
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot271  trt23
#> 2 subplot272  trt25
#> 3 subplot273  trt21
#> 4 subplot274  trt22
#> 5 subplot275  trt24
#> 
#> $mainplot56
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot276  trt23
#> 2 subplot277  trt21
#> 3 subplot278  trt25
#> 4 subplot279  trt24
#> 5 subplot280  trt22
#> 
#> $mainplot57
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot281  trt21
#> 2 subplot282  trt24
#> 3 subplot283  trt23
#> 4 subplot284  trt22
#> 5 subplot285  trt25
#> 
#> $mainplot58
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot286  trt21
#> 2 subplot287  trt24
#> 3 subplot288  trt23
#> 4 subplot289  trt22
#> 5 subplot290  trt25
#> 
#> $mainplot59
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot291  trt24
#> 2 subplot292  trt25
#> 3 subplot293  trt23
#> 4 subplot294  trt22
#> 5 subplot295  trt21
#> 
#> $mainplot60
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot296  trt22
#> 2 subplot297  trt25
#> 3 subplot298  trt23
#> 4 subplot299  trt21
#> 5 subplot300  trt24
#> 
#> $mainplot61
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot301  trt24
#> 2 subplot302  trt21
#> 3 subplot303  trt22
#> 4 subplot304  trt23
#> 5 subplot305  trt25
#> 
#> $mainplot62
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot306  trt24
#> 2 subplot307  trt21
#> 3 subplot308  trt23
#> 4 subplot309  trt25
#> 5 subplot310  trt22
#> 
#> $mainplot63
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot311  trt22
#> 2 subplot312  trt21
#> 3 subplot313  trt23
#> 4 subplot314  trt25
#> 5 subplot315  trt24
#> 
#> $mainplot64
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot316  trt24
#> 2 subplot317  trt25
#> 3 subplot318  trt21
#> 4 subplot319  trt23
#> 5 subplot320  trt22
#> 
#> $mainplot65
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot321  trt23
#> 2 subplot322  trt22
#> 3 subplot323  trt24
#> 4 subplot324  trt25
#> 5 subplot325  trt21
#> 
#> $mainplot66
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot326  trt25
#> 2 subplot327  trt24
#> 3 subplot328  trt22
#> 4 subplot329  trt23
#> 5 subplot330  trt21
#> 
#> $mainplot67
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot331  trt25
#> 2 subplot332  trt23
#> 3 subplot333  trt24
#> 4 subplot334  trt21
#> 5 subplot335  trt22
#> 
#> $mainplot68
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot336  trt22
#> 2 subplot337  trt24
#> 3 subplot338  trt23
#> 4 subplot339  trt25
#> 5 subplot340  trt21
#> 
#> $mainplot69
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot341  trt24
#> 2 subplot342  trt23
#> 3 subplot343  trt21
#> 4 subplot344  trt22
#> 5 subplot345  trt25
#> 
#> $mainplot70
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot346  trt21
#> 2 subplot347  trt25
#> 3 subplot348  trt24
#> 4 subplot349  trt22
#> 5 subplot350  trt23
#> 
#> $mainplot71
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot351  trt23
#> 2 subplot352  trt25
#> 3 subplot353  trt24
#> 4 subplot354  trt22
#> 5 subplot355  trt21
#> 
#> $mainplot72
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot356  trt24
#> 2 subplot357  trt23
#> 3 subplot358  trt22
#> 4 subplot359  trt25
#> 5 subplot360  trt21
#> 
#> $mainplot73
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot361  trt21
#> 2 subplot362  trt24
#> 3 subplot363  trt25
#> 4 subplot364  trt22
#> 5 subplot365  trt23
#> 
#> $mainplot74
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot366  trt25
#> 2 subplot367  trt24
#> 3 subplot368  trt22
#> 4 subplot369  trt23
#> 5 subplot370  trt21
#> 
#> $mainplot75
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot371  trt24
#> 2 subplot372  trt23
#> 3 subplot373  trt21
#> 4 subplot374  trt22
#> 5 subplot375  trt25
#> 
#> $mainplot76
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot376  trt25
#> 2 subplot377  trt23
#> 3 subplot378  trt21
#> 4 subplot379  trt22
#> 5 subplot380  trt24
#> 
#> $mainplot77
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot381  trt24
#> 2 subplot382  trt22
#> 3 subplot383  trt21
#> 4 subplot384  trt23
#> 5 subplot385  trt25
#> 
#> $mainplot78
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot386  trt22
#> 2 subplot387  trt25
#> 3 subplot388  trt21
#> 4 subplot389  trt24
#> 5 subplot390  trt23
#> 
#> $mainplot79
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot391  trt25
#> 2 subplot392  trt24
#> 3 subplot393  trt23
#> 4 subplot394  trt22
#> 5 subplot395  trt21
#> 
#> $mainplot80
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot396  trt22
#> 2 subplot397  trt23
#> 3 subplot398  trt24
#> 4 subplot399  trt21
#> 5 subplot400  trt25
#> 
#> $mainplot81
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot401  trt24
#> 2 subplot402  trt25
#> 3 subplot403  trt22
#> 4 subplot404  trt23
#> 5 subplot405  trt21
#> 
#> $mainplot82
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot406  trt24
#> 2 subplot407  trt22
#> 3 subplot408  trt23
#> 4 subplot409  trt25
#> 5 subplot410  trt21
#> 
#> $mainplot83
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot411  trt25
#> 2 subplot412  trt23
#> 3 subplot413  trt24
#> 4 subplot414  trt21
#> 5 subplot415  trt22
#> 
#> $mainplot84
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot416  trt25
#> 2 subplot417  trt22
#> 3 subplot418  trt24
#> 4 subplot419  trt21
#> 5 subplot420  trt23
#> 
#> $mainplot85
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot421  trt25
#> 2 subplot422  trt23
#> 3 subplot423  trt21
#> 4 subplot424  trt22
#> 5 subplot425  trt24
#> 
#> $mainplot86
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot426  trt25
#> 2 subplot427  trt21
#> 3 subplot428  trt22
#> 4 subplot429  trt23
#> 5 subplot430  trt24
#> 
#> $mainplot87
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot431  trt25
#> 2 subplot432  trt24
#> 3 subplot433  trt23
#> 4 subplot434  trt22
#> 5 subplot435  trt21
#> 
#> $mainplot88
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot436  trt24
#> 2 subplot437  trt21
#> 3 subplot438  trt23
#> 4 subplot439  trt22
#> 5 subplot440  trt25
#> 
#> $mainplot89
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot441  trt21
#> 2 subplot442  trt25
#> 3 subplot443  trt23
#> 4 subplot444  trt22
#> 5 subplot445  trt24
#> 
#> $mainplot90
#> # A tibble: 5 × 2
#>      subplot   trt2
#>     <U(450)> <T(5)>
#> 1 subplot446  trt25
#> 2 subplot447  trt21
#> 3 subplot448  trt23
#> 4 subplot449  trt22
#> 5 subplot450  trt24
#> 
spd %>% count_by(trt1)
#>    trt1 mainplot
#> 1 trt11       10
#> 2 trt12       10
#> 3 trt13       10
#> 4 trt14       10
#> 5 trt15       10
#> 6 trt16       10
#> 7 trt17       10
#> 8 trt18       10
#> 9 trt19       10

fac <- takeout(menu_factorial(trt = c(2, 2, 2)))
fac %>% count_by(where(~is_trt(.x)))
#>    trt1  trt2  trt3 unit
#> 1 trt11 trt21 trt31    5
#> 2 trt12 trt21 trt31    5
#> 3 trt11 trt22 trt31    5
#> 4 trt12 trt22 trt31    5
#> 5 trt11 trt21 trt32    5
#> 6 trt12 trt21 trt32    5
#> 7 trt11 trt22 trt32    5
#> 8 trt12 trt22 trt32    5