Split or count the data according to certain factors

This function has a similar result with split() where it returns a named list with names corresponding to the levels of the separating factor (or concatenated strings if multiple separating factors). The key differences to split(), are that the splitting factor does not appear in the elements of the list and only linked factors and their ancestors appear in the output, e.g. if treatment is applied to wholeplot and subplots are nested within subplots, then the subplot will not be shown in the output if split by treatment.

Usage

split_by(.data, ..., .sep = ":", .remove_empty = TRUE)

count_by(.data, ..., .remove_empty = TRUE)

Arguments

.data

An edibble table.

...

The factors to split or count by. You cannot split by a record factor or a factor that uniquely indexes the smallest unit in the design. You cannot also combine treatment and unit factors together.

.sep

The separator to use if more than one factor to split by.

.remove_empty

Remove empty combinations. Default is TRUE.

Value

A named list.

See also

Examples

spd <- takeout(menu_split())
spd %>% split(trt1)
#> Error: object 'trt1' not found
spd %>% split_by(trt1)
#> $trt11
#> # A tibble: 8 × 1
#>     mainplot
#>   <unit(64)>
#> 1 mainplot07
#> 2 mainplot08
#> 3 mainplot21
#> 4 mainplot22
#> 5 mainplot25
#> 6 mainplot28
#> 7 mainplot48
#> 8 mainplot52
#> 
#> $trt12
#> # A tibble: 8 × 1
#>     mainplot
#>   <unit(64)>
#> 1 mainplot02
#> 2 mainplot04
#> 3 mainplot20
#> 4 mainplot24
#> 5 mainplot33
#> 6 mainplot35
#> 7 mainplot40
#> 8 mainplot45
#> 
#> $trt13
#> # A tibble: 8 × 1
#>     mainplot
#>   <unit(64)>
#> 1 mainplot18
#> 2 mainplot19
#> 3 mainplot26
#> 4 mainplot27
#> 5 mainplot29
#> 6 mainplot36
#> 7 mainplot58
#> 8 mainplot59
#> 
#> $trt14
#> # A tibble: 8 × 1
#>     mainplot
#>   <unit(64)>
#> 1 mainplot01
#> 2 mainplot03
#> 3 mainplot09
#> 4 mainplot12
#> 5 mainplot32
#> 6 mainplot37
#> 7 mainplot46
#> 8 mainplot47
#> 
#> $trt15
#> # A tibble: 8 × 1
#>     mainplot
#>   <unit(64)>
#> 1 mainplot11
#> 2 mainplot14
#> 3 mainplot34
#> 4 mainplot38
#> 5 mainplot51
#> 6 mainplot53
#> 7 mainplot54
#> 8 mainplot64
#> 
#> $trt16
#> # A tibble: 8 × 1
#>     mainplot
#>   <unit(64)>
#> 1 mainplot15
#> 2 mainplot30
#> 3 mainplot44
#> 4 mainplot55
#> 5 mainplot56
#> 6 mainplot57
#> 7 mainplot61
#> 8 mainplot63
#> 
#> $trt17
#> # A tibble: 8 × 1
#>     mainplot
#>   <unit(64)>
#> 1 mainplot10
#> 2 mainplot13
#> 3 mainplot16
#> 4 mainplot17
#> 5 mainplot23
#> 6 mainplot39
#> 7 mainplot43
#> 8 mainplot62
#> 
#> $trt18
#> # A tibble: 8 × 1
#>     mainplot
#>   <unit(64)>
#> 1 mainplot05
#> 2 mainplot06
#> 3 mainplot31
#> 4 mainplot41
#> 5 mainplot42
#> 6 mainplot49
#> 7 mainplot50
#> 8 mainplot60
#> 
spd %>% split_by(trt2)
#> $trt21
#> # A tibble: 64 × 2
#>        subplot   mainplot
#>    <unit(576)> <unit(64)>
#>  1  subplot009 mainplot01
#>  2  subplot010 mainplot02
#>  3  subplot026 mainplot03
#>  4  subplot030 mainplot04
#>  5  subplot041 mainplot05
#>  6  subplot052 mainplot06
#>  7  subplot059 mainplot07
#>  8  subplot071 mainplot08
#>  9  subplot080 mainplot09
#> 10  subplot085 mainplot10
#> # ℹ 54 more rows
#> 
#> $trt22
#> # A tibble: 64 × 2
#>        subplot   mainplot
#>    <unit(576)> <unit(64)>
#>  1  subplot005 mainplot01
#>  2  subplot013 mainplot02
#>  3  subplot020 mainplot03
#>  4  subplot031 mainplot04
#>  5  subplot044 mainplot05
#>  6  subplot053 mainplot06
#>  7  subplot058 mainplot07
#>  8  subplot066 mainplot08
#>  9  subplot076 mainplot09
#> 10  subplot088 mainplot10
#> # ℹ 54 more rows
#> 
#> $trt23
#> # A tibble: 64 × 2
#>        subplot   mainplot
#>    <unit(576)> <unit(64)>
#>  1  subplot001 mainplot01
#>  2  subplot014 mainplot02
#>  3  subplot021 mainplot03
#>  4  subplot032 mainplot04
#>  5  subplot045 mainplot05
#>  6  subplot047 mainplot06
#>  7  subplot057 mainplot07
#>  8  subplot065 mainplot08
#>  9  subplot077 mainplot09
#> 10  subplot084 mainplot10
#> # ℹ 54 more rows
#> 
#> $trt24
#> # A tibble: 64 × 2
#>        subplot   mainplot
#>    <unit(576)> <unit(64)>
#>  1  subplot002 mainplot01
#>  2  subplot011 mainplot02
#>  3  subplot023 mainplot03
#>  4  subplot034 mainplot04
#>  5  subplot037 mainplot05
#>  6  subplot050 mainplot06
#>  7  subplot055 mainplot07
#>  8  subplot070 mainplot08
#>  9  subplot075 mainplot09
#> 10  subplot086 mainplot10
#> # ℹ 54 more rows
#> 
#> $trt25
#> # A tibble: 64 × 2
#>        subplot   mainplot
#>    <unit(576)> <unit(64)>
#>  1  subplot003 mainplot01
#>  2  subplot017 mainplot02
#>  3  subplot019 mainplot03
#>  4  subplot028 mainplot04
#>  5  subplot043 mainplot05
#>  6  subplot048 mainplot06
#>  7  subplot063 mainplot07
#>  8  subplot072 mainplot08
#>  9  subplot078 mainplot09
#> 10  subplot089 mainplot10
#> # ℹ 54 more rows
#> 
#> $trt26
#> # A tibble: 64 × 2
#>        subplot   mainplot
#>    <unit(576)> <unit(64)>
#>  1  subplot006 mainplot01
#>  2  subplot016 mainplot02
#>  3  subplot027 mainplot03
#>  4  subplot033 mainplot04
#>  5  subplot039 mainplot05
#>  6  subplot054 mainplot06
#>  7  subplot060 mainplot07
#>  8  subplot068 mainplot08
#>  9  subplot073 mainplot09
#> 10  subplot083 mainplot10
#> # ℹ 54 more rows
#> 
#> $trt27
#> # A tibble: 64 × 2
#>        subplot   mainplot
#>    <unit(576)> <unit(64)>
#>  1  subplot007 mainplot01
#>  2  subplot015 mainplot02
#>  3  subplot024 mainplot03
#>  4  subplot029 mainplot04
#>  5  subplot042 mainplot05
#>  6  subplot051 mainplot06
#>  7  subplot062 mainplot07
#>  8  subplot067 mainplot08
#>  9  subplot074 mainplot09
#> 10  subplot090 mainplot10
#> # ℹ 54 more rows
#> 
#> $trt28
#> # A tibble: 64 × 2
#>        subplot   mainplot
#>    <unit(576)> <unit(64)>
#>  1  subplot004 mainplot01
#>  2  subplot018 mainplot02
#>  3  subplot022 mainplot03
#>  4  subplot035 mainplot04
#>  5  subplot040 mainplot05
#>  6  subplot046 mainplot06
#>  7  subplot061 mainplot07
#>  8  subplot064 mainplot08
#>  9  subplot079 mainplot09
#> 10  subplot087 mainplot10
#> # ℹ 54 more rows
#> 
#> $trt29
#> # A tibble: 64 × 2
#>        subplot   mainplot
#>    <unit(576)> <unit(64)>
#>  1  subplot008 mainplot01
#>  2  subplot012 mainplot02
#>  3  subplot025 mainplot03
#>  4  subplot036 mainplot04
#>  5  subplot038 mainplot05
#>  6  subplot049 mainplot06
#>  7  subplot056 mainplot07
#>  8  subplot069 mainplot08
#>  9  subplot081 mainplot09
#> 10  subplot082 mainplot10
#> # ℹ 54 more rows
#> 
spd %>% split_by(mainplot)
#> $mainplot01
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot001    trt23
#> 2  subplot002    trt24
#> 3  subplot003    trt25
#> 4  subplot004    trt28
#> 5  subplot005    trt22
#> 6  subplot006    trt26
#> 7  subplot007    trt27
#> 8  subplot008    trt29
#> 9  subplot009    trt21
#> 
#> $mainplot02
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot010    trt21
#> 2  subplot011    trt24
#> 3  subplot012    trt29
#> 4  subplot013    trt22
#> 5  subplot014    trt23
#> 6  subplot015    trt27
#> 7  subplot016    trt26
#> 8  subplot017    trt25
#> 9  subplot018    trt28
#> 
#> $mainplot03
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot019    trt25
#> 2  subplot020    trt22
#> 3  subplot021    trt23
#> 4  subplot022    trt28
#> 5  subplot023    trt24
#> 6  subplot024    trt27
#> 7  subplot025    trt29
#> 8  subplot026    trt21
#> 9  subplot027    trt26
#> 
#> $mainplot04
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot028    trt25
#> 2  subplot029    trt27
#> 3  subplot030    trt21
#> 4  subplot031    trt22
#> 5  subplot032    trt23
#> 6  subplot033    trt26
#> 7  subplot034    trt24
#> 8  subplot035    trt28
#> 9  subplot036    trt29
#> 
#> $mainplot05
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot037    trt24
#> 2  subplot038    trt29
#> 3  subplot039    trt26
#> 4  subplot040    trt28
#> 5  subplot041    trt21
#> 6  subplot042    trt27
#> 7  subplot043    trt25
#> 8  subplot044    trt22
#> 9  subplot045    trt23
#> 
#> $mainplot06
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot046    trt28
#> 2  subplot047    trt23
#> 3  subplot048    trt25
#> 4  subplot049    trt29
#> 5  subplot050    trt24
#> 6  subplot051    trt27
#> 7  subplot052    trt21
#> 8  subplot053    trt22
#> 9  subplot054    trt26
#> 
#> $mainplot07
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot055    trt24
#> 2  subplot056    trt29
#> 3  subplot057    trt23
#> 4  subplot058    trt22
#> 5  subplot059    trt21
#> 6  subplot060    trt26
#> 7  subplot061    trt28
#> 8  subplot062    trt27
#> 9  subplot063    trt25
#> 
#> $mainplot08
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot064    trt28
#> 2  subplot065    trt23
#> 3  subplot066    trt22
#> 4  subplot067    trt27
#> 5  subplot068    trt26
#> 6  subplot069    trt29
#> 7  subplot070    trt24
#> 8  subplot071    trt21
#> 9  subplot072    trt25
#> 
#> $mainplot09
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot073    trt26
#> 2  subplot074    trt27
#> 3  subplot075    trt24
#> 4  subplot076    trt22
#> 5  subplot077    trt23
#> 6  subplot078    trt25
#> 7  subplot079    trt28
#> 8  subplot080    trt21
#> 9  subplot081    trt29
#> 
#> $mainplot10
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot082    trt29
#> 2  subplot083    trt26
#> 3  subplot084    trt23
#> 4  subplot085    trt21
#> 5  subplot086    trt24
#> 6  subplot087    trt28
#> 7  subplot088    trt22
#> 8  subplot089    trt25
#> 9  subplot090    trt27
#> 
#> $mainplot11
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot091    trt24
#> 2  subplot092    trt29
#> 3  subplot093    trt27
#> 4  subplot094    trt28
#> 5  subplot095    trt25
#> 6  subplot096    trt21
#> 7  subplot097    trt23
#> 8  subplot098    trt26
#> 9  subplot099    trt22
#> 
#> $mainplot12
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot100    trt24
#> 2  subplot101    trt25
#> 3  subplot102    trt28
#> 4  subplot103    trt22
#> 5  subplot104    trt23
#> 6  subplot105    trt21
#> 7  subplot106    trt29
#> 8  subplot107    trt26
#> 9  subplot108    trt27
#> 
#> $mainplot13
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot109    trt21
#> 2  subplot110    trt24
#> 3  subplot111    trt28
#> 4  subplot112    trt25
#> 5  subplot113    trt27
#> 6  subplot114    trt22
#> 7  subplot115    trt26
#> 8  subplot116    trt29
#> 9  subplot117    trt23
#> 
#> $mainplot14
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot118    trt29
#> 2  subplot119    trt25
#> 3  subplot120    trt23
#> 4  subplot121    trt24
#> 5  subplot122    trt22
#> 6  subplot123    trt28
#> 7  subplot124    trt26
#> 8  subplot125    trt27
#> 9  subplot126    trt21
#> 
#> $mainplot15
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot127    trt24
#> 2  subplot128    trt28
#> 3  subplot129    trt27
#> 4  subplot130    trt29
#> 5  subplot131    trt21
#> 6  subplot132    trt23
#> 7  subplot133    trt22
#> 8  subplot134    trt26
#> 9  subplot135    trt25
#> 
#> $mainplot16
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot136    trt22
#> 2  subplot137    trt29
#> 3  subplot138    trt26
#> 4  subplot139    trt21
#> 5  subplot140    trt23
#> 6  subplot141    trt27
#> 7  subplot142    trt24
#> 8  subplot143    trt25
#> 9  subplot144    trt28
#> 
#> $mainplot17
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot145    trt27
#> 2  subplot146    trt26
#> 3  subplot147    trt25
#> 4  subplot148    trt23
#> 5  subplot149    trt28
#> 6  subplot150    trt21
#> 7  subplot151    trt22
#> 8  subplot152    trt24
#> 9  subplot153    trt29
#> 
#> $mainplot18
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot154    trt24
#> 2  subplot155    trt21
#> 3  subplot156    trt27
#> 4  subplot157    trt25
#> 5  subplot158    trt29
#> 6  subplot159    trt28
#> 7  subplot160    trt22
#> 8  subplot161    trt23
#> 9  subplot162    trt26
#> 
#> $mainplot19
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot163    trt21
#> 2  subplot164    trt22
#> 3  subplot165    trt23
#> 4  subplot166    trt24
#> 5  subplot167    trt25
#> 6  subplot168    trt28
#> 7  subplot169    trt29
#> 8  subplot170    trt27
#> 9  subplot171    trt26
#> 
#> $mainplot20
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot172    trt25
#> 2  subplot173    trt22
#> 3  subplot174    trt28
#> 4  subplot175    trt26
#> 5  subplot176    trt21
#> 6  subplot177    trt29
#> 7  subplot178    trt27
#> 8  subplot179    trt23
#> 9  subplot180    trt24
#> 
#> $mainplot21
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot181    trt29
#> 2  subplot182    trt24
#> 3  subplot183    trt25
#> 4  subplot184    trt28
#> 5  subplot185    trt23
#> 6  subplot186    trt21
#> 7  subplot187    trt26
#> 8  subplot188    trt22
#> 9  subplot189    trt27
#> 
#> $mainplot22
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot190    trt26
#> 2  subplot191    trt23
#> 3  subplot192    trt27
#> 4  subplot193    trt29
#> 5  subplot194    trt22
#> 6  subplot195    trt21
#> 7  subplot196    trt28
#> 8  subplot197    trt24
#> 9  subplot198    trt25
#> 
#> $mainplot23
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot199    trt23
#> 2  subplot200    trt22
#> 3  subplot201    trt29
#> 4  subplot202    trt25
#> 5  subplot203    trt27
#> 6  subplot204    trt28
#> 7  subplot205    trt24
#> 8  subplot206    trt26
#> 9  subplot207    trt21
#> 
#> $mainplot24
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot208    trt25
#> 2  subplot209    trt27
#> 3  subplot210    trt26
#> 4  subplot211    trt28
#> 5  subplot212    trt29
#> 6  subplot213    trt21
#> 7  subplot214    trt23
#> 8  subplot215    trt22
#> 9  subplot216    trt24
#> 
#> $mainplot25
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot217    trt24
#> 2  subplot218    trt27
#> 3  subplot219    trt29
#> 4  subplot220    trt28
#> 5  subplot221    trt21
#> 6  subplot222    trt25
#> 7  subplot223    trt22
#> 8  subplot224    trt23
#> 9  subplot225    trt26
#> 
#> $mainplot26
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot226    trt24
#> 2  subplot227    trt23
#> 3  subplot228    trt22
#> 4  subplot229    trt21
#> 5  subplot230    trt27
#> 6  subplot231    trt26
#> 7  subplot232    trt29
#> 8  subplot233    trt25
#> 9  subplot234    trt28
#> 
#> $mainplot27
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot235    trt28
#> 2  subplot236    trt21
#> 3  subplot237    trt25
#> 4  subplot238    trt24
#> 5  subplot239    trt23
#> 6  subplot240    trt27
#> 7  subplot241    trt26
#> 8  subplot242    trt22
#> 9  subplot243    trt29
#> 
#> $mainplot28
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot244    trt27
#> 2  subplot245    trt25
#> 3  subplot246    trt21
#> 4  subplot247    trt24
#> 5  subplot248    trt29
#> 6  subplot249    trt22
#> 7  subplot250    trt26
#> 8  subplot251    trt23
#> 9  subplot252    trt28
#> 
#> $mainplot29
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot253    trt24
#> 2  subplot254    trt23
#> 3  subplot255    trt27
#> 4  subplot256    trt29
#> 5  subplot257    trt28
#> 6  subplot258    trt22
#> 7  subplot259    trt25
#> 8  subplot260    trt26
#> 9  subplot261    trt21
#> 
#> $mainplot30
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot262    trt29
#> 2  subplot263    trt26
#> 3  subplot264    trt28
#> 4  subplot265    trt27
#> 5  subplot266    trt25
#> 6  subplot267    trt21
#> 7  subplot268    trt24
#> 8  subplot269    trt22
#> 9  subplot270    trt23
#> 
#> $mainplot31
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot271    trt26
#> 2  subplot272    trt24
#> 3  subplot273    trt29
#> 4  subplot274    trt21
#> 5  subplot275    trt28
#> 6  subplot276    trt23
#> 7  subplot277    trt27
#> 8  subplot278    trt25
#> 9  subplot279    trt22
#> 
#> $mainplot32
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot280    trt24
#> 2  subplot281    trt27
#> 3  subplot282    trt22
#> 4  subplot283    trt25
#> 5  subplot284    trt28
#> 6  subplot285    trt21
#> 7  subplot286    trt29
#> 8  subplot287    trt23
#> 9  subplot288    trt26
#> 
#> $mainplot33
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot289    trt22
#> 2  subplot290    trt23
#> 3  subplot291    trt24
#> 4  subplot292    trt28
#> 5  subplot293    trt21
#> 6  subplot294    trt29
#> 7  subplot295    trt27
#> 8  subplot296    trt26
#> 9  subplot297    trt25
#> 
#> $mainplot34
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot298    trt25
#> 2  subplot299    trt21
#> 3  subplot300    trt28
#> 4  subplot301    trt26
#> 5  subplot302    trt23
#> 6  subplot303    trt24
#> 7  subplot304    trt27
#> 8  subplot305    trt29
#> 9  subplot306    trt22
#> 
#> $mainplot35
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot307    trt24
#> 2  subplot308    trt27
#> 3  subplot309    trt22
#> 4  subplot310    trt28
#> 5  subplot311    trt23
#> 6  subplot312    trt29
#> 7  subplot313    trt21
#> 8  subplot314    trt25
#> 9  subplot315    trt26
#> 
#> $mainplot36
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot316    trt25
#> 2  subplot317    trt28
#> 3  subplot318    trt22
#> 4  subplot319    trt26
#> 5  subplot320    trt29
#> 6  subplot321    trt23
#> 7  subplot322    trt21
#> 8  subplot323    trt24
#> 9  subplot324    trt27
#> 
#> $mainplot37
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot325    trt29
#> 2  subplot326    trt24
#> 3  subplot327    trt22
#> 4  subplot328    trt21
#> 5  subplot329    trt27
#> 6  subplot330    trt23
#> 7  subplot331    trt26
#> 8  subplot332    trt25
#> 9  subplot333    trt28
#> 
#> $mainplot38
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot334    trt23
#> 2  subplot335    trt25
#> 3  subplot336    trt21
#> 4  subplot337    trt22
#> 5  subplot338    trt24
#> 6  subplot339    trt28
#> 7  subplot340    trt27
#> 8  subplot341    trt26
#> 9  subplot342    trt29
#> 
#> $mainplot39
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot343    trt28
#> 2  subplot344    trt29
#> 3  subplot345    trt23
#> 4  subplot346    trt27
#> 5  subplot347    trt25
#> 6  subplot348    trt24
#> 7  subplot349    trt22
#> 8  subplot350    trt26
#> 9  subplot351    trt21
#> 
#> $mainplot40
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot352    trt29
#> 2  subplot353    trt21
#> 3  subplot354    trt22
#> 4  subplot355    trt26
#> 5  subplot356    trt27
#> 6  subplot357    trt25
#> 7  subplot358    trt23
#> 8  subplot359    trt24
#> 9  subplot360    trt28
#> 
#> $mainplot41
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot361    trt29
#> 2  subplot362    trt26
#> 3  subplot363    trt22
#> 4  subplot364    trt27
#> 5  subplot365    trt23
#> 6  subplot366    trt25
#> 7  subplot367    trt21
#> 8  subplot368    trt28
#> 9  subplot369    trt24
#> 
#> $mainplot42
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot370    trt27
#> 2  subplot371    trt26
#> 3  subplot372    trt28
#> 4  subplot373    trt22
#> 5  subplot374    trt29
#> 6  subplot375    trt25
#> 7  subplot376    trt21
#> 8  subplot377    trt24
#> 9  subplot378    trt23
#> 
#> $mainplot43
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot379    trt28
#> 2  subplot380    trt22
#> 3  subplot381    trt27
#> 4  subplot382    trt24
#> 5  subplot383    trt29
#> 6  subplot384    trt25
#> 7  subplot385    trt26
#> 8  subplot386    trt23
#> 9  subplot387    trt21
#> 
#> $mainplot44
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot388    trt26
#> 2  subplot389    trt27
#> 3  subplot390    trt21
#> 4  subplot391    trt29
#> 5  subplot392    trt23
#> 6  subplot393    trt24
#> 7  subplot394    trt28
#> 8  subplot395    trt25
#> 9  subplot396    trt22
#> 
#> $mainplot45
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot397    trt21
#> 2  subplot398    trt25
#> 3  subplot399    trt26
#> 4  subplot400    trt23
#> 5  subplot401    trt28
#> 6  subplot402    trt24
#> 7  subplot403    trt29
#> 8  subplot404    trt22
#> 9  subplot405    trt27
#> 
#> $mainplot46
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot406    trt26
#> 2  subplot407    trt23
#> 3  subplot408    trt27
#> 4  subplot409    trt29
#> 5  subplot410    trt25
#> 6  subplot411    trt21
#> 7  subplot412    trt24
#> 8  subplot413    trt28
#> 9  subplot414    trt22
#> 
#> $mainplot47
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot415    trt26
#> 2  subplot416    trt29
#> 3  subplot417    trt25
#> 4  subplot418    trt22
#> 5  subplot419    trt21
#> 6  subplot420    trt24
#> 7  subplot421    trt23
#> 8  subplot422    trt28
#> 9  subplot423    trt27
#> 
#> $mainplot48
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot424    trt21
#> 2  subplot425    trt25
#> 3  subplot426    trt28
#> 4  subplot427    trt22
#> 5  subplot428    trt23
#> 6  subplot429    trt27
#> 7  subplot430    trt24
#> 8  subplot431    trt26
#> 9  subplot432    trt29
#> 
#> $mainplot49
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot433    trt28
#> 2  subplot434    trt29
#> 3  subplot435    trt23
#> 4  subplot436    trt25
#> 5  subplot437    trt21
#> 6  subplot438    trt26
#> 7  subplot439    trt22
#> 8  subplot440    trt27
#> 9  subplot441    trt24
#> 
#> $mainplot50
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot442    trt26
#> 2  subplot443    trt22
#> 3  subplot444    trt23
#> 4  subplot445    trt25
#> 5  subplot446    trt28
#> 6  subplot447    trt27
#> 7  subplot448    trt29
#> 8  subplot449    trt24
#> 9  subplot450    trt21
#> 
#> $mainplot51
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot451    trt27
#> 2  subplot452    trt26
#> 3  subplot453    trt22
#> 4  subplot454    trt21
#> 5  subplot455    trt24
#> 6  subplot456    trt23
#> 7  subplot457    trt28
#> 8  subplot458    trt29
#> 9  subplot459    trt25
#> 
#> $mainplot52
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot460    trt21
#> 2  subplot461    trt23
#> 3  subplot462    trt22
#> 4  subplot463    trt24
#> 5  subplot464    trt25
#> 6  subplot465    trt27
#> 7  subplot466    trt28
#> 8  subplot467    trt29
#> 9  subplot468    trt26
#> 
#> $mainplot53
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot469    trt21
#> 2  subplot470    trt23
#> 3  subplot471    trt25
#> 4  subplot472    trt28
#> 5  subplot473    trt26
#> 6  subplot474    trt22
#> 7  subplot475    trt24
#> 8  subplot476    trt29
#> 9  subplot477    trt27
#> 
#> $mainplot54
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot478    trt21
#> 2  subplot479    trt24
#> 3  subplot480    trt28
#> 4  subplot481    trt22
#> 5  subplot482    trt29
#> 6  subplot483    trt27
#> 7  subplot484    trt23
#> 8  subplot485    trt25
#> 9  subplot486    trt26
#> 
#> $mainplot55
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot487    trt21
#> 2  subplot488    trt24
#> 3  subplot489    trt28
#> 4  subplot490    trt23
#> 5  subplot491    trt25
#> 6  subplot492    trt29
#> 7  subplot493    trt26
#> 8  subplot494    trt22
#> 9  subplot495    trt27
#> 
#> $mainplot56
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot496    trt26
#> 2  subplot497    trt22
#> 3  subplot498    trt25
#> 4  subplot499    trt28
#> 5  subplot500    trt24
#> 6  subplot501    trt23
#> 7  subplot502    trt27
#> 8  subplot503    trt29
#> 9  subplot504    trt21
#> 
#> $mainplot57
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot505    trt21
#> 2  subplot506    trt27
#> 3  subplot507    trt25
#> 4  subplot508    trt22
#> 5  subplot509    trt28
#> 6  subplot510    trt24
#> 7  subplot511    trt26
#> 8  subplot512    trt29
#> 9  subplot513    trt23
#> 
#> $mainplot58
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot514    trt29
#> 2  subplot515    trt21
#> 3  subplot516    trt25
#> 4  subplot517    trt24
#> 5  subplot518    trt26
#> 6  subplot519    trt23
#> 7  subplot520    trt28
#> 8  subplot521    trt22
#> 9  subplot522    trt27
#> 
#> $mainplot59
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot523    trt27
#> 2  subplot524    trt26
#> 3  subplot525    trt24
#> 4  subplot526    trt23
#> 5  subplot527    trt21
#> 6  subplot528    trt29
#> 7  subplot529    trt28
#> 8  subplot530    trt25
#> 9  subplot531    trt22
#> 
#> $mainplot60
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot532    trt22
#> 2  subplot533    trt26
#> 3  subplot534    trt24
#> 4  subplot535    trt28
#> 5  subplot536    trt29
#> 6  subplot537    trt25
#> 7  subplot538    trt27
#> 8  subplot539    trt21
#> 9  subplot540    trt23
#> 
#> $mainplot61
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot541    trt22
#> 2  subplot542    trt25
#> 3  subplot543    trt21
#> 4  subplot544    trt29
#> 5  subplot545    trt24
#> 6  subplot546    trt28
#> 7  subplot547    trt23
#> 8  subplot548    trt26
#> 9  subplot549    trt27
#> 
#> $mainplot62
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot550    trt28
#> 2  subplot551    trt27
#> 3  subplot552    trt23
#> 4  subplot553    trt21
#> 5  subplot554    trt29
#> 6  subplot555    trt24
#> 7  subplot556    trt25
#> 8  subplot557    trt22
#> 9  subplot558    trt26
#> 
#> $mainplot63
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot559    trt24
#> 2  subplot560    trt23
#> 3  subplot561    trt22
#> 4  subplot562    trt27
#> 5  subplot563    trt29
#> 6  subplot564    trt25
#> 7  subplot565    trt26
#> 8  subplot566    trt21
#> 9  subplot567    trt28
#> 
#> $mainplot64
#> # A tibble: 9 × 2
#>       subplot     trt2
#>   <unit(576)> <trt(9)>
#> 1  subplot568    trt29
#> 2  subplot569    trt26
#> 3  subplot570    trt27
#> 4  subplot571    trt28
#> 5  subplot572    trt24
#> 6  subplot573    trt22
#> 7  subplot574    trt25
#> 8  subplot575    trt21
#> 9  subplot576    trt23
#> 
spd %>% count_by(trt1)
#>    trt1 mainplot
#> 1 trt11        8
#> 2 trt12        8
#> 3 trt13        8
#> 4 trt14        8
#> 5 trt15        8
#> 6 trt16        8
#> 7 trt17        8
#> 8 trt18        8

fac <- takeout(menu_factorial(trt = c(2, 2, 2)))
fac %>% count_by(where(~is_trt(.x)))
#>    trt1  trt2  trt3 unit
#> 1 trt11 trt21 trt31    2
#> 2 trt12 trt21 trt31    2
#> 3 trt11 trt22 trt31    2
#> 4 trt12 trt22 trt31    2
#> 5 trt11 trt21 trt32    2
#> 6 trt12 trt21 trt32    2
#> 7 trt11 trt22 trt32    2
#> 8 trt12 trt22 trt32    2