Split or count the data according to certain factors

This function has a similar result with split() where it returns a named list with names corresponding to the levels of the separating factor (or concatenated strings if multiple separating factors). The key differences to split(), are that the splitting factor does not appear in the elements of the list and only linked factors and their ancestors appear in the output, e.g. if treatment is applied to wholeplot and subplots are nested within subplots, then the subplot will not be shown in the output if split by treatment.

Usage

split_by(.data, ..., .sep = ":", .remove_empty = TRUE)

count_by(.data, ..., .remove_empty = TRUE)

Arguments

.data

An edibble table.

...

The factors to split or count by. You cannot split by a record factor or a factor that uniquely indexes the smallest unit in the design. You cannot also combine treatment and unit factors together.

.sep

The separator to use if more than one factor to split by.

.remove_empty

Remove empty combinations. Default is TRUE.

Value

A named list.

See also

Examples

spd <- takeout(menu_split())
spd %>% split(trt1)
#> Error: object 'trt1' not found
spd %>% split_by(trt1)
#> $trt11
#> # A tibble: 6 × 1
#>     mainplot
#>   <unit(54)>
#> 1 mainplot20
#> 2 mainplot24
#> 3 mainplot34
#> 4 mainplot37
#> 5 mainplot43
#> 6 mainplot54
#> 
#> $trt12
#> # A tibble: 6 × 1
#>     mainplot
#>   <unit(54)>
#> 1 mainplot09
#> 2 mainplot12
#> 3 mainplot31
#> 4 mainplot35
#> 5 mainplot41
#> 6 mainplot49
#> 
#> $trt13
#> # A tibble: 6 × 1
#>     mainplot
#>   <unit(54)>
#> 1 mainplot08
#> 2 mainplot13
#> 3 mainplot14
#> 4 mainplot29
#> 5 mainplot33
#> 6 mainplot46
#> 
#> $trt14
#> # A tibble: 6 × 1
#>     mainplot
#>   <unit(54)>
#> 1 mainplot03
#> 2 mainplot11
#> 3 mainplot18
#> 4 mainplot30
#> 5 mainplot39
#> 6 mainplot45
#> 
#> $trt15
#> # A tibble: 6 × 1
#>     mainplot
#>   <unit(54)>
#> 1 mainplot15
#> 2 mainplot16
#> 3 mainplot28
#> 4 mainplot32
#> 5 mainplot38
#> 6 mainplot52
#> 
#> $trt16
#> # A tibble: 6 × 1
#>     mainplot
#>   <unit(54)>
#> 1 mainplot06
#> 2 mainplot21
#> 3 mainplot22
#> 4 mainplot26
#> 5 mainplot27
#> 6 mainplot51
#> 
#> $trt17
#> # A tibble: 6 × 1
#>     mainplot
#>   <unit(54)>
#> 1 mainplot02
#> 2 mainplot05
#> 3 mainplot07
#> 4 mainplot10
#> 5 mainplot44
#> 6 mainplot47
#> 
#> $trt18
#> # A tibble: 6 × 1
#>     mainplot
#>   <unit(54)>
#> 1 mainplot04
#> 2 mainplot23
#> 3 mainplot42
#> 4 mainplot48
#> 5 mainplot50
#> 6 mainplot53
#> 
#> $trt19
#> # A tibble: 6 × 1
#>     mainplot
#>   <unit(54)>
#> 1 mainplot01
#> 2 mainplot17
#> 3 mainplot19
#> 4 mainplot25
#> 5 mainplot36
#> 6 mainplot40
#> 
spd %>% split_by(trt2)
#> $trt21
#> # A tibble: 54 × 2
#>        subplot   mainplot
#>    <unit(432)> <unit(54)>
#>  1  subplot005 mainplot01
#>  2  subplot012 mainplot02
#>  3  subplot019 mainplot03
#>  4  subplot027 mainplot04
#>  5  subplot033 mainplot05
#>  6  subplot047 mainplot06
#>  7  subplot051 mainplot07
#>  8  subplot059 mainplot08
#>  9  subplot068 mainplot09
#> 10  subplot074 mainplot10
#> # ℹ 44 more rows
#> 
#> $trt22
#> # A tibble: 54 × 2
#>        subplot   mainplot
#>    <unit(432)> <unit(54)>
#>  1  subplot003 mainplot01
#>  2  subplot009 mainplot02
#>  3  subplot023 mainplot03
#>  4  subplot032 mainplot04
#>  5  subplot036 mainplot05
#>  6  subplot041 mainplot06
#>  7  subplot052 mainplot07
#>  8  subplot058 mainplot08
#>  9  subplot070 mainplot09
#> 10  subplot079 mainplot10
#> # ℹ 44 more rows
#> 
#> $trt23
#> # A tibble: 54 × 2
#>        subplot   mainplot
#>    <unit(432)> <unit(54)>
#>  1  subplot006 mainplot01
#>  2  subplot014 mainplot02
#>  3  subplot024 mainplot03
#>  4  subplot028 mainplot04
#>  5  subplot038 mainplot05
#>  6  subplot045 mainplot06
#>  7  subplot049 mainplot07
#>  8  subplot061 mainplot08
#>  9  subplot069 mainplot09
#> 10  subplot080 mainplot10
#> # ℹ 44 more rows
#> 
#> $trt24
#> # A tibble: 54 × 2
#>        subplot   mainplot
#>    <unit(432)> <unit(54)>
#>  1  subplot008 mainplot01
#>  2  subplot011 mainplot02
#>  3  subplot022 mainplot03
#>  4  subplot026 mainplot04
#>  5  subplot034 mainplot05
#>  6  subplot048 mainplot06
#>  7  subplot050 mainplot07
#>  8  subplot063 mainplot08
#>  9  subplot067 mainplot09
#> 10  subplot076 mainplot10
#> # ℹ 44 more rows
#> 
#> $trt25
#> # A tibble: 54 × 2
#>        subplot   mainplot
#>    <unit(432)> <unit(54)>
#>  1  subplot002 mainplot01
#>  2  subplot015 mainplot02
#>  3  subplot018 mainplot03
#>  4  subplot030 mainplot04
#>  5  subplot035 mainplot05
#>  6  subplot043 mainplot06
#>  7  subplot054 mainplot07
#>  8  subplot064 mainplot08
#>  9  subplot071 mainplot09
#> 10  subplot075 mainplot10
#> # ℹ 44 more rows
#> 
#> $trt26
#> # A tibble: 54 × 2
#>        subplot   mainplot
#>    <unit(432)> <unit(54)>
#>  1  subplot001 mainplot01
#>  2  subplot016 mainplot02
#>  3  subplot020 mainplot03
#>  4  subplot029 mainplot04
#>  5  subplot039 mainplot05
#>  6  subplot046 mainplot06
#>  7  subplot056 mainplot07
#>  8  subplot060 mainplot08
#>  9  subplot065 mainplot09
#> 10  subplot078 mainplot10
#> # ℹ 44 more rows
#> 
#> $trt27
#> # A tibble: 54 × 2
#>        subplot   mainplot
#>    <unit(432)> <unit(54)>
#>  1  subplot007 mainplot01
#>  2  subplot013 mainplot02
#>  3  subplot017 mainplot03
#>  4  subplot031 mainplot04
#>  5  subplot037 mainplot05
#>  6  subplot044 mainplot06
#>  7  subplot055 mainplot07
#>  8  subplot057 mainplot08
#>  9  subplot072 mainplot09
#> 10  subplot073 mainplot10
#> # ℹ 44 more rows
#> 
#> $trt28
#> # A tibble: 54 × 2
#>        subplot   mainplot
#>    <unit(432)> <unit(54)>
#>  1  subplot004 mainplot01
#>  2  subplot010 mainplot02
#>  3  subplot021 mainplot03
#>  4  subplot025 mainplot04
#>  5  subplot040 mainplot05
#>  6  subplot042 mainplot06
#>  7  subplot053 mainplot07
#>  8  subplot062 mainplot08
#>  9  subplot066 mainplot09
#> 10  subplot077 mainplot10
#> # ℹ 44 more rows
#> 
spd %>% split_by(mainplot)
#> $mainplot01
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot001    trt26
#> 2  subplot002    trt25
#> 3  subplot003    trt22
#> 4  subplot004    trt28
#> 5  subplot005    trt21
#> 6  subplot006    trt23
#> 7  subplot007    trt27
#> 8  subplot008    trt24
#> 
#> $mainplot02
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot009    trt22
#> 2  subplot010    trt28
#> 3  subplot011    trt24
#> 4  subplot012    trt21
#> 5  subplot013    trt27
#> 6  subplot014    trt23
#> 7  subplot015    trt25
#> 8  subplot016    trt26
#> 
#> $mainplot03
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot017    trt27
#> 2  subplot018    trt25
#> 3  subplot019    trt21
#> 4  subplot020    trt26
#> 5  subplot021    trt28
#> 6  subplot022    trt24
#> 7  subplot023    trt22
#> 8  subplot024    trt23
#> 
#> $mainplot04
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot025    trt28
#> 2  subplot026    trt24
#> 3  subplot027    trt21
#> 4  subplot028    trt23
#> 5  subplot029    trt26
#> 6  subplot030    trt25
#> 7  subplot031    trt27
#> 8  subplot032    trt22
#> 
#> $mainplot05
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot033    trt21
#> 2  subplot034    trt24
#> 3  subplot035    trt25
#> 4  subplot036    trt22
#> 5  subplot037    trt27
#> 6  subplot038    trt23
#> 7  subplot039    trt26
#> 8  subplot040    trt28
#> 
#> $mainplot06
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot041    trt22
#> 2  subplot042    trt28
#> 3  subplot043    trt25
#> 4  subplot044    trt27
#> 5  subplot045    trt23
#> 6  subplot046    trt26
#> 7  subplot047    trt21
#> 8  subplot048    trt24
#> 
#> $mainplot07
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot049    trt23
#> 2  subplot050    trt24
#> 3  subplot051    trt21
#> 4  subplot052    trt22
#> 5  subplot053    trt28
#> 6  subplot054    trt25
#> 7  subplot055    trt27
#> 8  subplot056    trt26
#> 
#> $mainplot08
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot057    trt27
#> 2  subplot058    trt22
#> 3  subplot059    trt21
#> 4  subplot060    trt26
#> 5  subplot061    trt23
#> 6  subplot062    trt28
#> 7  subplot063    trt24
#> 8  subplot064    trt25
#> 
#> $mainplot09
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot065    trt26
#> 2  subplot066    trt28
#> 3  subplot067    trt24
#> 4  subplot068    trt21
#> 5  subplot069    trt23
#> 6  subplot070    trt22
#> 7  subplot071    trt25
#> 8  subplot072    trt27
#> 
#> $mainplot10
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot073    trt27
#> 2  subplot074    trt21
#> 3  subplot075    trt25
#> 4  subplot076    trt24
#> 5  subplot077    trt28
#> 6  subplot078    trt26
#> 7  subplot079    trt22
#> 8  subplot080    trt23
#> 
#> $mainplot11
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot081    trt21
#> 2  subplot082    trt25
#> 3  subplot083    trt26
#> 4  subplot084    trt23
#> 5  subplot085    trt24
#> 6  subplot086    trt22
#> 7  subplot087    trt28
#> 8  subplot088    trt27
#> 
#> $mainplot12
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot089    trt24
#> 2  subplot090    trt26
#> 3  subplot091    trt27
#> 4  subplot092    trt22
#> 5  subplot093    trt28
#> 6  subplot094    trt23
#> 7  subplot095    trt25
#> 8  subplot096    trt21
#> 
#> $mainplot13
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot097    trt21
#> 2  subplot098    trt28
#> 3  subplot099    trt23
#> 4  subplot100    trt24
#> 5  subplot101    trt22
#> 6  subplot102    trt27
#> 7  subplot103    trt26
#> 8  subplot104    trt25
#> 
#> $mainplot14
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot105    trt24
#> 2  subplot106    trt27
#> 3  subplot107    trt28
#> 4  subplot108    trt23
#> 5  subplot109    trt25
#> 6  subplot110    trt22
#> 7  subplot111    trt26
#> 8  subplot112    trt21
#> 
#> $mainplot15
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot113    trt27
#> 2  subplot114    trt22
#> 3  subplot115    trt23
#> 4  subplot116    trt28
#> 5  subplot117    trt21
#> 6  subplot118    trt25
#> 7  subplot119    trt24
#> 8  subplot120    trt26
#> 
#> $mainplot16
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot121    trt22
#> 2  subplot122    trt21
#> 3  subplot123    trt23
#> 4  subplot124    trt25
#> 5  subplot125    trt28
#> 6  subplot126    trt24
#> 7  subplot127    trt27
#> 8  subplot128    trt26
#> 
#> $mainplot17
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot129    trt25
#> 2  subplot130    trt22
#> 3  subplot131    trt24
#> 4  subplot132    trt28
#> 5  subplot133    trt26
#> 6  subplot134    trt23
#> 7  subplot135    trt21
#> 8  subplot136    trt27
#> 
#> $mainplot18
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot137    trt22
#> 2  subplot138    trt21
#> 3  subplot139    trt25
#> 4  subplot140    trt24
#> 5  subplot141    trt26
#> 6  subplot142    trt27
#> 7  subplot143    trt28
#> 8  subplot144    trt23
#> 
#> $mainplot19
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot145    trt25
#> 2  subplot146    trt28
#> 3  subplot147    trt22
#> 4  subplot148    trt23
#> 5  subplot149    trt24
#> 6  subplot150    trt27
#> 7  subplot151    trt26
#> 8  subplot152    trt21
#> 
#> $mainplot20
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot153    trt27
#> 2  subplot154    trt24
#> 3  subplot155    trt21
#> 4  subplot156    trt22
#> 5  subplot157    trt26
#> 6  subplot158    trt25
#> 7  subplot159    trt23
#> 8  subplot160    trt28
#> 
#> $mainplot21
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot161    trt23
#> 2  subplot162    trt27
#> 3  subplot163    trt25
#> 4  subplot164    trt22
#> 5  subplot165    trt26
#> 6  subplot166    trt24
#> 7  subplot167    trt28
#> 8  subplot168    trt21
#> 
#> $mainplot22
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot169    trt25
#> 2  subplot170    trt28
#> 3  subplot171    trt21
#> 4  subplot172    trt23
#> 5  subplot173    trt27
#> 6  subplot174    trt26
#> 7  subplot175    trt24
#> 8  subplot176    trt22
#> 
#> $mainplot23
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot177    trt27
#> 2  subplot178    trt26
#> 3  subplot179    trt25
#> 4  subplot180    trt21
#> 5  subplot181    trt23
#> 6  subplot182    trt24
#> 7  subplot183    trt28
#> 8  subplot184    trt22
#> 
#> $mainplot24
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot185    trt21
#> 2  subplot186    trt26
#> 3  subplot187    trt22
#> 4  subplot188    trt28
#> 5  subplot189    trt25
#> 6  subplot190    trt27
#> 7  subplot191    trt23
#> 8  subplot192    trt24
#> 
#> $mainplot25
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot193    trt23
#> 2  subplot194    trt26
#> 3  subplot195    trt21
#> 4  subplot196    trt27
#> 5  subplot197    trt28
#> 6  subplot198    trt25
#> 7  subplot199    trt24
#> 8  subplot200    trt22
#> 
#> $mainplot26
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot201    trt27
#> 2  subplot202    trt26
#> 3  subplot203    trt28
#> 4  subplot204    trt23
#> 5  subplot205    trt21
#> 6  subplot206    trt25
#> 7  subplot207    trt24
#> 8  subplot208    trt22
#> 
#> $mainplot27
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot209    trt25
#> 2  subplot210    trt26
#> 3  subplot211    trt22
#> 4  subplot212    trt28
#> 5  subplot213    trt27
#> 6  subplot214    trt21
#> 7  subplot215    trt23
#> 8  subplot216    trt24
#> 
#> $mainplot28
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot217    trt27
#> 2  subplot218    trt24
#> 3  subplot219    trt21
#> 4  subplot220    trt26
#> 5  subplot221    trt23
#> 6  subplot222    trt22
#> 7  subplot223    trt25
#> 8  subplot224    trt28
#> 
#> $mainplot29
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot225    trt26
#> 2  subplot226    trt28
#> 3  subplot227    trt21
#> 4  subplot228    trt25
#> 5  subplot229    trt23
#> 6  subplot230    trt22
#> 7  subplot231    trt27
#> 8  subplot232    trt24
#> 
#> $mainplot30
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot233    trt26
#> 2  subplot234    trt25
#> 3  subplot235    trt28
#> 4  subplot236    trt24
#> 5  subplot237    trt22
#> 6  subplot238    trt27
#> 7  subplot239    trt23
#> 8  subplot240    trt21
#> 
#> $mainplot31
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot241    trt22
#> 2  subplot242    trt26
#> 3  subplot243    trt21
#> 4  subplot244    trt25
#> 5  subplot245    trt24
#> 6  subplot246    trt28
#> 7  subplot247    trt27
#> 8  subplot248    trt23
#> 
#> $mainplot32
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot249    trt23
#> 2  subplot250    trt24
#> 3  subplot251    trt26
#> 4  subplot252    trt21
#> 5  subplot253    trt22
#> 6  subplot254    trt25
#> 7  subplot255    trt28
#> 8  subplot256    trt27
#> 
#> $mainplot33
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot257    trt25
#> 2  subplot258    trt26
#> 3  subplot259    trt27
#> 4  subplot260    trt28
#> 5  subplot261    trt21
#> 6  subplot262    trt22
#> 7  subplot263    trt24
#> 8  subplot264    trt23
#> 
#> $mainplot34
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot265    trt21
#> 2  subplot266    trt23
#> 3  subplot267    trt28
#> 4  subplot268    trt22
#> 5  subplot269    trt25
#> 6  subplot270    trt24
#> 7  subplot271    trt27
#> 8  subplot272    trt26
#> 
#> $mainplot35
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot273    trt25
#> 2  subplot274    trt24
#> 3  subplot275    trt23
#> 4  subplot276    trt26
#> 5  subplot277    trt27
#> 6  subplot278    trt22
#> 7  subplot279    trt21
#> 8  subplot280    trt28
#> 
#> $mainplot36
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot281    trt22
#> 2  subplot282    trt23
#> 3  subplot283    trt25
#> 4  subplot284    trt28
#> 5  subplot285    trt24
#> 6  subplot286    trt26
#> 7  subplot287    trt21
#> 8  subplot288    trt27
#> 
#> $mainplot37
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot289    trt24
#> 2  subplot290    trt21
#> 3  subplot291    trt26
#> 4  subplot292    trt27
#> 5  subplot293    trt22
#> 6  subplot294    trt23
#> 7  subplot295    trt25
#> 8  subplot296    trt28
#> 
#> $mainplot38
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot297    trt28
#> 2  subplot298    trt27
#> 3  subplot299    trt26
#> 4  subplot300    trt23
#> 5  subplot301    trt25
#> 6  subplot302    trt24
#> 7  subplot303    trt22
#> 8  subplot304    trt21
#> 
#> $mainplot39
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot305    trt25
#> 2  subplot306    trt26
#> 3  subplot307    trt22
#> 4  subplot308    trt27
#> 5  subplot309    trt28
#> 6  subplot310    trt23
#> 7  subplot311    trt21
#> 8  subplot312    trt24
#> 
#> $mainplot40
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot313    trt22
#> 2  subplot314    trt21
#> 3  subplot315    trt23
#> 4  subplot316    trt24
#> 5  subplot317    trt25
#> 6  subplot318    trt27
#> 7  subplot319    trt28
#> 8  subplot320    trt26
#> 
#> $mainplot41
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot321    trt22
#> 2  subplot322    trt26
#> 3  subplot323    trt25
#> 4  subplot324    trt21
#> 5  subplot325    trt23
#> 6  subplot326    trt27
#> 7  subplot327    trt24
#> 8  subplot328    trt28
#> 
#> $mainplot42
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot329    trt25
#> 2  subplot330    trt27
#> 3  subplot331    trt21
#> 4  subplot332    trt24
#> 5  subplot333    trt28
#> 6  subplot334    trt26
#> 7  subplot335    trt22
#> 8  subplot336    trt23
#> 
#> $mainplot43
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot337    trt21
#> 2  subplot338    trt24
#> 3  subplot339    trt25
#> 4  subplot340    trt26
#> 5  subplot341    trt28
#> 6  subplot342    trt23
#> 7  subplot343    trt27
#> 8  subplot344    trt22
#> 
#> $mainplot44
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot345    trt21
#> 2  subplot346    trt28
#> 3  subplot347    trt27
#> 4  subplot348    trt25
#> 5  subplot349    trt24
#> 6  subplot350    trt26
#> 7  subplot351    trt22
#> 8  subplot352    trt23
#> 
#> $mainplot45
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot353    trt28
#> 2  subplot354    trt21
#> 3  subplot355    trt24
#> 4  subplot356    trt27
#> 5  subplot357    trt25
#> 6  subplot358    trt22
#> 7  subplot359    trt23
#> 8  subplot360    trt26
#> 
#> $mainplot46
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot361    trt24
#> 2  subplot362    trt22
#> 3  subplot363    trt25
#> 4  subplot364    trt26
#> 5  subplot365    trt21
#> 6  subplot366    trt28
#> 7  subplot367    trt27
#> 8  subplot368    trt23
#> 
#> $mainplot47
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot369    trt27
#> 2  subplot370    trt22
#> 3  subplot371    trt25
#> 4  subplot372    trt21
#> 5  subplot373    trt23
#> 6  subplot374    trt24
#> 7  subplot375    trt26
#> 8  subplot376    trt28
#> 
#> $mainplot48
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot377    trt27
#> 2  subplot378    trt23
#> 3  subplot379    trt21
#> 4  subplot380    trt28
#> 5  subplot381    trt26
#> 6  subplot382    trt22
#> 7  subplot383    trt25
#> 8  subplot384    trt24
#> 
#> $mainplot49
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot385    trt24
#> 2  subplot386    trt21
#> 3  subplot387    trt27
#> 4  subplot388    trt25
#> 5  subplot389    trt22
#> 6  subplot390    trt23
#> 7  subplot391    trt26
#> 8  subplot392    trt28
#> 
#> $mainplot50
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot393    trt24
#> 2  subplot394    trt22
#> 3  subplot395    trt25
#> 4  subplot396    trt28
#> 5  subplot397    trt23
#> 6  subplot398    trt26
#> 7  subplot399    trt27
#> 8  subplot400    trt21
#> 
#> $mainplot51
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot401    trt25
#> 2  subplot402    trt21
#> 3  subplot403    trt27
#> 4  subplot404    trt23
#> 5  subplot405    trt28
#> 6  subplot406    trt22
#> 7  subplot407    trt24
#> 8  subplot408    trt26
#> 
#> $mainplot52
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot409    trt23
#> 2  subplot410    trt27
#> 3  subplot411    trt25
#> 4  subplot412    trt28
#> 5  subplot413    trt26
#> 6  subplot414    trt24
#> 7  subplot415    trt22
#> 8  subplot416    trt21
#> 
#> $mainplot53
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot417    trt26
#> 2  subplot418    trt23
#> 3  subplot419    trt27
#> 4  subplot420    trt28
#> 5  subplot421    trt24
#> 6  subplot422    trt25
#> 7  subplot423    trt21
#> 8  subplot424    trt22
#> 
#> $mainplot54
#> # A tibble: 8 × 2
#>       subplot     trt2
#>   <unit(432)> <trt(8)>
#> 1  subplot425    trt25
#> 2  subplot426    trt24
#> 3  subplot427    trt23
#> 4  subplot428    trt28
#> 5  subplot429    trt27
#> 6  subplot430    trt21
#> 7  subplot431    trt22
#> 8  subplot432    trt26
#> 
spd %>% count_by(trt1)
#>    trt1 mainplot
#> 1 trt11        6
#> 2 trt12        6
#> 3 trt13        6
#> 4 trt14        6
#> 5 trt15        6
#> 6 trt16        6
#> 7 trt17        6
#> 8 trt18        6
#> 9 trt19        6

fac <- takeout(menu_factorial(trt = c(2, 2, 2)))
fac %>% count_by(where(~is_trt(.x)))
#>    trt1  trt2  trt3 unit
#> 1 trt11 trt21 trt31    6
#> 2 trt12 trt21 trt31    6
#> 3 trt11 trt22 trt31    6
#> 4 trt12 trt22 trt31    6
#> 5 trt11 trt21 trt32    6
#> 6 trt12 trt21 trt32    6
#> 7 trt11 trt22 trt32    6
#> 8 trt12 trt22 trt32    6